Split kirra_verifier_service.rs handlers into per-domain modules
The service binary had grown to 6324 lines: ~50 axum route handlers
interleaved with middleware, startup sentinel, helpers, main/build_app
and 14 test modules in one file.

Convert it to a directory-backed binary: the route handler functions move
verbatim into ten sibling submodules under
src/bin/kirra_verifier_service/, declared with #[path] (a flat bin file is
a crate root, so its modules resolve as siblings without the attribute):

  attestation, fleet, audit, action_filter, industrial,
  federation, actuator, fabric, console, operators

It's a pure move. To minimise blast radius on a security-critical binary,
ONLY free handler fns move; every struct/enum/DTO, all shared helpers
(require_admin_token, gate_posture, supervisor-key checks, the systemd
watchdog, …), the startup sentinel, main, build_app and ALL test modules
stay in the root file. Consequences, all mechanical:

  - moved handlers are widened from `fn`/`async fn` to `pub(crate)` so the
    root can re-export them (`use <module>::*`); build_app and the in-file
    tests therefore reference them unqualified, unchanged.
  - each submodule pulls the root's DTOs, helpers and `use` imports via
    `use super::*` (descendant-module visibility); handlers read DTO fields
    (defined in the ancestor root) without any field-visibility change.
  - console_html's `include_str!` path gains one `../` (it sits one
    directory deeper now).
  - the P1 store_offload_guard test now `include_str!`s all ten submodules
    plus the root and scans each, so a worker-pinning `save_posture_event_
    chained` write reintroduced in any handler file is still caught.

A sorted-line diff against the original confirms every other line is
byte-identical: the only deltas are the 53 `pub(crate)` promotions, the
module scaffolding, and the guard rewrite. Route wiring, auth layering and
all CLAUDE.md security invariants are untouched.

736 root-crate tests pass (incl. the 73 in-binary tests); clippy clean.
