+++ src/bin/kirra_verifier_service.rs
@@ -55,6 +55,42 @@ use kirra_verifier::fabric::router::FabricRouter;
 use kirra_verifier::fabric::telemetry::FabricTelemetry;
 use kirra_verifier::fabric::causal_log::FabricCausalLog;
 
+// Route handlers, split by domain into sibling submodules. Each holds
+// `pub(crate)` handler fns that share the binary's helpers, DTOs and `use`
+// imports via `use super::*` (descendant-module visibility). Re-exported
+// below so `build_app` and the in-file tests reference them unqualified.
+#[path = "kirra_verifier_service/attestation.rs"]
+mod attestation;
+#[path = "kirra_verifier_service/fleet.rs"]
+mod fleet;
+#[path = "kirra_verifier_service/audit.rs"]
+mod audit;
+#[path = "kirra_verifier_service/action_filter.rs"]
+mod action_filter;
+#[path = "kirra_verifier_service/industrial.rs"]
+mod industrial;
+#[path = "kirra_verifier_service/federation.rs"]
+mod federation;
+#[path = "kirra_verifier_service/actuator.rs"]
+mod actuator;
+#[path = "kirra_verifier_service/fabric.rs"]
+mod fabric;
+#[path = "kirra_verifier_service/console.rs"]
+mod console;
+#[path = "kirra_verifier_service/operators.rs"]
+mod operators;
+use attestation::*;
+use fleet::*;
+use audit::*;
+use action_filter::*;
+use industrial::*;
+use federation::*;
+use actuator::*;
+use fabric::*;
+use console::*;
+use operators::*;
+
+
 // --- Auth middleware ---------------------------------------------------------
 
 async fn require_admin_token(request: Request, next: Next) -> Result<Response, StatusCode> {
@@ -402,23 +438,6 @@ fn emit_posture_event(state: &AppState, event_type: &str, node_id: Option<String
     });
 }
 
-async fn system_posture_stream(
-    State(svc): State<Arc<ServiceState>>,
-) -> Sse<impl tokio_stream::Stream<Item = Result<Event, Infallible>>> {
-    let rx = svc.app.posture_tx.subscribe();
-    let stream = BroadcastStream::new(rx).filter_map(|msg| {
-        match msg {
-            Ok(event) => serde_json::to_string(&event).ok().map(|data| {
-                Ok(Event::default().data(data))
-            }),
-            Err(tokio_stream::wrappers::errors::BroadcastStreamRecvError::Lagged(n)) => {
-                tracing::warn!(skipped = n, "posture stream subscriber lagged; frames dropped");
-                None
-            }
-        }
-    });
-    Sse::new(stream).keep_alive(KeepAlive::default())
-}
 
 // --- Request / response types -----------------------------------------------
 
@@ -505,19 +524,7 @@ struct RegisterAvAssetRequest {
 
 // --- Handlers ----------------------------------------------------------------
 
-async fn health() -> Json<HealthResponse> {
-    Json(HealthResponse { status: "ok".to_string() })
-}
 
-async fn ready(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    match svc.app.store.with(|store| store.health_check()) {
-        Ok(()) => (StatusCode::OK, Json(HealthResponse { status: "ready".to_string() }))
-            .into_response(),
-        Err(_) => (StatusCode::SERVICE_UNAVAILABLE,
-                   Json(HealthResponse { status: "db_unavailable".to_string() }))
-            .into_response(),
-    }
-}
 
 /// Outcome of the offloaded federation commit (`submit_federated_report`), mapped
 /// to an HTTP response on the async side. Side effects that must happen on the
@@ -533,291 +540,12 @@ enum FedCommitOutcome {
     Fenced(String),
 }
 
-async fn export_backup(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    let exported_at_ms = now_ms();
-    // Three full-table loads (nodes + dependencies + all posture events) — the
-    // heaviest read. `call_read` runs it off the worker pool AND against a
-    // read-only replica connection, so a large dump neither pins a tokio worker
-    // nor contends the writer mutex (a concurrent write proceeds unblocked).
-    let result = svc.app.store.call_read(move |store| {
-        let nodes = store.load_nodes().ok()?;
-        let dependencies = store.load_dependencies().ok()?;
-        let posture_events = store.load_all_posture_events().ok()?;
-        Some(BackupExport { exported_at_ms, nodes, dependencies, posture_events })
-    })
-    .await;
-    match result {
-        Ok(Some(export)) => Json(export).into_response(),
-        Ok(None) => (StatusCode::INTERNAL_SERVER_ERROR,
-                     Json(json!({ "error": "failed to export backup" }))).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "store lock poisoned" }))).into_response(),
-    }
-}
 
-async fn register_node(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<RegisterNodeRequest>,
-) -> impl IntoResponse {
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    let now = now_ms();
-    let node = RegisteredNode {
-        node_id: req.node_id.clone(),
-        status: NodeTrustState::Unknown,
-        registered_at_ms: now,
-        last_trust_update_ms: 0,
-        ak_public_pem: req.ak_public_pem,
-        expected_pcr16_digest_hex: req.expected_pcr16_digest_hex,
-        site: req.site,
-        firmware_version: req.firmware_version,
-    };
 
-    // TPM-quote policy (#572 follow-up) is committed BEFORE the node record, so
-    // a node that requires a hardware quote is never live without its policy
-    // (fail-closed: no window where the requirement silently does not apply). A
-    // store error here fails the whole registration — the node is not inserted.
-    {
-        let policy_err = svc.app.store.with(|store| {
-            store.set_node_attestation_policy(&req.node_id, req.require_tpm_quote).is_err()
-        });
-        if policy_err {
-            return (StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(json!({ "error": "failed to persist attestation policy" }))).into_response();
-        }
-    }
 
-    if svc.app.persist_and_insert_node(node).is_err() {
-        return (StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": "failed to persist node" }))).into_response();
-    }
 
-    (StatusCode::CREATED, Json(json!({ "node_id": req.node_id, "status": "registered" }))).into_response()
-}
 
-async fn issue_challenge(
-    State(svc): State<Arc<ServiceState>>,
-    Path(node_id): Path<String>,
-) -> impl IntoResponse {
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    if !svc.app.nodes.contains_key(&node_id) {
-        return (StatusCode::NOT_FOUND,
-                Json(json!({ "error": "node not registered" }))).into_response();
-    }
-    // #147: the challenge nonce comes from a CSPRNG (OsRng), NEVER the wall
-    // clock. A `SystemTime`-derived nonce is predictable and can collide within
-    // a single nanosecond; single-use + TTL + node-binding are enforced by the
-    // challenge store and the verify-then-consume order in `verify_attestation`.
-    let nonce = kirra_verifier::verifier::generate_challenge_nonce();
-    svc.app.issue_challenge(&node_id, nonce, now_ms());
-    (StatusCode::OK, Json(json!({ "node_id": node_id, "nonce": nonce }))).into_response()
-}
 
-async fn verify_attestation(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<VerifyAttestationRequest>,
-) -> impl IntoResponse {
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    let now = now_ms();
-
-    // SAFETY: SG9 | REQ: attestation-node-proven-identity | TEST: valid_signature_verifies,legacy_admin_token_hmac_proof_is_rejected,absent_registered_key_fails_closed,wrong_key_is_rejected
-    // (#73) Node-PROVEN identity: the node must prove possession of the
-    // PRIVATE attestation key matching the `ak_public_pem` it registered, by
-    // signing the (node_id, nonce) challenge with Ed25519. The prior
-    // `HMAC(KIRRA_ADMIN_TOKEN, nonce)` proof was admin-ASSERTED trust —
-    // anyone with the admin token could attest any node. Fail-closed: a node
-    // with no registered AK, a malformed key, a malformed proof, or a bad
-    // signature is rejected here, before the nonce is consumed or any trust
-    // state is written. PCR16 (measured-boot) binding: when the node registered an
-    // `expected_pcr16_digest_hex`, the proof must carry a matching digest BOUND
-    // into the AK signature (`verify_attestation_proof_with_pcr16`); a node with no
-    // expectation is unaffected. A hardware TPM *quote* (the deeper measured-boot
-    // root) is enforced just below for a node whose policy requires it.
-    let (ak_public_pem, expected_pcr16) = match svc.app.nodes.get(&req.node_id) {
-        Some(node) => (node.ak_public_pem.clone(), node.expected_pcr16_digest_hex.clone()),
-        None => return (StatusCode::NOT_FOUND,
-                        Json(json!({ "error": "node not registered" }))).into_response(),
-    };
-
-    if let Err(reason) = kirra_verifier::attestation::verify_attestation_proof_with_pcr16(
-        ak_public_pem.as_deref(),
-        &req.node_id,
-        req.nonce,
-        &req.proof_hex,
-        expected_pcr16.as_deref(),
-        req.presented_pcr16_digest_hex.as_deref(),
-    ) {
-        // No registered key is a precondition failure (403); a measured-boot
-        // mismatch is a forbidden boot state (403); a present-but-failing
-        // signature is an authentication failure (401). Either way the
-        // attestation is REFUSED — never accepted by default, and the nonce is
-        // NOT consumed (this is before `consume_challenge`), so a node can retry
-        // with a corrected measured boot.
-        use kirra_verifier::attestation::AttestationError;
-        let status = match reason {
-            AttestationError::NoRegisteredKey | AttestationError::Pcr16Mismatch => {
-                StatusCode::FORBIDDEN
-            }
-            _ => StatusCode::UNAUTHORIZED,
-        };
-        tracing::warn!(node_id = %req.node_id, reason = %reason.as_str(),
-            "attestation proof rejected (fail-closed, #73)");
-        return (status, Json(json!({ "error": reason.as_str() }))).into_response();
-    }
-
-    // SAFETY: SG9 | REQ: attestation-tpm-quote-enforcement | TEST: tpm_quote_required_but_absent_is_refused,tpm_quote_valid_attests_node_trusted,tpm_quote_invalid_is_refused_and_nonce_preserved,tpm_quote_policy_absent_is_back_compat
-    // Hardware-rooted measured boot. The #73/#572 check above proves a
-    // SELF-REPORTED PCR16 digest under the AK — a node in control of its AK
-    // could sign a FALSE digest. A node enrolled with `require_tpm_quote` must
-    // additionally present a TPM QUOTE: the TPM itself signs the live PCR bank +
-    // the challenge nonce, so a forged boot state cannot be minted in software.
-    // Fail-closed: a policy-lookup error, a required-but-absent quote, an absent
-    // expectation to check against, or an invalid quote all REJECT. Runs BEFORE
-    // `consume_challenge`, so a quote failure does NOT burn the nonce (retry).
-    let require_quote = match svc.app.store.with(|store| store.node_requires_tpm_quote(&req.node_id)) {
-        Ok(v) => v,
-        Err(_) => return (StatusCode::INTERNAL_SERVER_ERROR,
-                          Json(json!({ "error": "attestation policy lookup failed" }))).into_response(),
-    };
-    match (&req.tpm_quote, require_quote) {
-        (Some(quote), _) => {
-            // The quote attests a HASH OVER the PCR16 value; the registered datum
-            // is the value itself. Bridge via `expected_single_pcr_digest_hex`.
-            let expected_digest = match expected_pcr16
-                .as_deref()
-                .and_then(kirra_verifier::tpm_quote::expected_single_pcr_digest_hex)
-            {
-                Some(d) => d,
-                None => {
-                    tracing::warn!(node_id = %req.node_id,
-                        "tpm quote presented but node has no expected PCR16 to verify against");
-                    return (StatusCode::FORBIDDEN,
-                            Json(json!({ "error": "tpm quote presented but no expected PCR16 registered" }))).into_response();
-                }
-            };
-            let nonce_bytes = req.nonce.to_be_bytes();
-            if let Err(e) = kirra_verifier::tpm_quote::verify_tpm_quote(
-                ak_public_pem.as_deref(),
-                &nonce_bytes,
-                &expected_digest,
-                &quote.quote_msg_hex,
-                &quote.signature_hex,
-            ) {
-                use kirra_verifier::tpm_quote::TpmQuoteError;
-                // A bad signature / unparseable bytes is an authentication failure
-                // (401); everything else (no key, wrong magic/type, nonce, PCR
-                // selection, digest) is a forbidden boot/identity state (403).
-                let status = match e {
-                    TpmQuoteError::SignatureInvalid
-                    | TpmQuoteError::MalformedEncoding
-                    | TpmQuoteError::MalformedQuote => StatusCode::UNAUTHORIZED,
-                    _ => StatusCode::FORBIDDEN,
-                };
-                tracing::warn!(node_id = %req.node_id, reason = %e.as_str(),
-                    "tpm quote rejected (fail-closed) — nonce preserved");
-                return (status, Json(json!({ "error": e.as_str() }))).into_response();
-            }
-        }
-        (None, true) => {
-            tracing::warn!(node_id = %req.node_id,
-                "node policy requires a tpm quote but none was presented");
-            return (StatusCode::FORBIDDEN,
-                    Json(json!({ "error": "tpm quote required by node policy but not presented" }))).into_response();
-        }
-        (None, false) => { /* back-compat: no quote required, none presented */ }
-    }
-
-    if !svc.app.consume_challenge(&req.node_id, req.nonce, now) {
-        return (StatusCode::CONFLICT,
-                Json(json!({ "error": "nonce absent, expired, or already consumed" }))).into_response();
-    }
-
-    let updated = match svc.app.nodes.get(&req.node_id) {
-        Some(existing) => RegisteredNode {
-            node_id: existing.node_id.clone(),
-            status: NodeTrustState::Trusted,
-            registered_at_ms: existing.registered_at_ms,
-            last_trust_update_ms: now,
-            ak_public_pem: existing.ak_public_pem.clone(),
-            expected_pcr16_digest_hex: existing.expected_pcr16_digest_hex.clone(),
-            site: existing.site.clone(),
-            firmware_version: existing.firmware_version.clone(),
-        },
-        None => return (StatusCode::NOT_FOUND,
-                        Json(json!({ "error": "node not registered" }))).into_response(),
-    };
-
-    if svc.app.persist_and_insert_node(updated).is_err() {
-        return (StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": "failed to persist trust state" }))).into_response();
-    }
-
-    let posture = svc.app.calculate_posture(&req.node_id);
-    if let Ok(posture_json) = serde_json::to_string(&posture) {
-        // P1: durable audit write off the worker pool. Own the node id (reused below).
-        let node_id_c = req.node_id.clone();
-        let _ = svc.app.store.call(move |store| {
-            if let Err(e) = store.save_posture_event_chained(
-                &node_id_c, "ATTESTATION_TRUSTED", &posture_json, None, now,
-            ) {
-                tracing::error!(error=%e, node_id=%node_id_c,
-                    "AUDIT-CHAIN WRITE FAILED for ATTESTATION_TRUSTED — event missing from tamper-evident log");
-            }
-        }).await;
-    }
-    emit_posture_event(&svc.app, "NODE_STATUS_CHANGED", Some(req.node_id.clone()));
-    enqueue_recalc(&svc, kirra_verifier::posture_engine_v2::PostureRecalcTrigger::NodeTrustChanged {
-        node_id: req.node_id.clone(),
-        reason:  "ATTESTATION_TRUSTED".to_string(),
-    });
-
-    (StatusCode::OK, Json(json!({ "node_id": req.node_id, "attested": true }))).into_response()
-}
-
-async fn get_node_status(
-    State(svc): State<Arc<ServiceState>>,
-    Path(node_id): Path<String>,
-) -> impl IntoResponse {
-    match svc.app.nodes.get(&node_id) {
-        Some(node) => {
-            let status = match &node.status {
-                NodeTrustState::Trusted => "Trusted",
-                NodeTrustState::Untrusted(_) => "Untrusted",
-                NodeTrustState::Unknown => "Unknown",
-            };
-            (StatusCode::OK, Json(AttestationStatusResponse {
-                node_id: node_id.clone(),
-                status: status.to_string(),
-                registered_at_ms: node.registered_at_ms,
-            })).into_response()
-        }
-        None => (StatusCode::NOT_FOUND, Json(json!({ "error": "not found" }))).into_response(),
-    }
-}
-
-async fn get_fleet_posture(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    let postures: Vec<FleetNodePosture> = svc.app.nodes
-        .iter()
-        .map(|entry| svc.app.calculate_posture(entry.key()))
-        .collect();
-    Json(json!({ "fleet": postures }))
-}
-
-async fn get_node_posture(
-    State(svc): State<Arc<ServiceState>>,
-    Path(node_id): Path<String>,
-) -> impl IntoResponse {
-    let posture = svc.app.calculate_posture(&node_id);
-    Json(posture)
-}
 
 #[derive(Serialize)]
 struct AvSubsystemView {
@@ -830,30 +558,6 @@ struct AvSubsystemView {
     recovery_streak_start_ms: u64,
 }
 
-/// Read-only listing of registered AV subsystem diagnostics (confidence floor,
-/// recovery streak, last telemetry). Admin-gated; no secrets returned. (#385)
-async fn list_av_subsystems(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    match svc.app.store.with(|store| store.load_av_subsystems()) {
-        Ok(rows) => {
-            let subsystems: Vec<AvSubsystemView> = rows
-                .into_iter()
-                .map(|r| AvSubsystemView {
-                    node_id: r.node_id,
-                    subsystem_type: r.subsystem_type,
-                    hardware_id: r.hardware_id,
-                    confidence_floor: r.confidence_floor,
-                    last_telemetry_ms: r.last_telemetry_ms,
-                    recovery_streak_count: r.recovery_streak_count,
-                    recovery_streak_start_ms: r.recovery_streak_start_ms,
-                })
-                .collect();
-            let total = subsystems.len();
-            Json(json!({ "subsystems": subsystems, "total": total })).into_response()
-        }
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to load av subsystems" }))).into_response(),
-    }
-}
 
 #[derive(Serialize)]
 struct OperatorView {
@@ -864,134 +568,10 @@ struct OperatorView {
     active: bool,
 }
 
-/// Read-only listing of registered operators. Admin-gated. Exposes only the
-/// public-key FINGERPRINT (never the PEM), matching the write-side convention. (#385)
-async fn list_operators(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    match svc.app.store.with(|store| store.load_operators()) {
-        Ok(rows) => {
-            let operators: Vec<OperatorView> = rows
-                .into_iter()
-                .map(|r| {
-                    let active = r.is_active();
-                    OperatorView {
-                        operator_key_fingerprint:
-                            kirra_verifier::attestation::operator_key_fingerprint(&r.pubkey_pem)
-                                .unwrap_or_else(|| "unparseable".to_string()),
-                        operator_id: r.operator_id,
-                        registered_at_ms: r.registered_at_ms,
-                        revoked_at_ms: r.revoked_at_ms,
-                        active,
-                    }
-                })
-                .collect();
-            let total = operators.len();
-            Json(json!({ "operators": operators, "total": total })).into_response()
-        }
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to load operators" }))).into_response(),
-    }
-}
 
-async fn register_dependencies(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<RegisterDependenciesRequest>,
-) -> impl IntoResponse {
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    if svc.app.persist_and_insert_deps(&req.node_id, req.depends_on).is_err() {
-        return (StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": "failed to persist dependencies" }))).into_response();
-    }
 
-    let posture = svc.app.calculate_posture(&req.node_id);
-    let now = now_ms();
-    if let Ok(posture_json) = serde_json::to_string(&posture) {
-        // P1: durable audit write off the worker pool. Own the node id (reused in
-        // the response); the posture json is owned and moves in.
-        let node_id_c = req.node_id.clone();
-        let _ = svc.app.store.call(move |store| {
-            if let Err(e) = store.save_posture_event_chained(
-                &node_id_c, "DEPENDENCY_UPDATED", &posture_json, None, now,
-            ) {
-                tracing::error!(error=%e, node_id=%node_id_c,
-                    "AUDIT-CHAIN WRITE FAILED for DEPENDENCY_UPDATED — event missing from tamper-evident log");
-            }
-        }).await;
-    }
-    emit_posture_event(&svc.app, "DEPENDENCY_GRAPH_MUTATED", Some(req.node_id.clone()));
-    enqueue_recalc(&svc, kirra_verifier::posture_engine_v2::PostureRecalcTrigger::DependencyGraphChanged);
 
-    (StatusCode::OK, Json(json!({ "node_id": req.node_id, "dependencies_registered": true }))).into_response()
-}
 
-async fn get_node_history(
-    State(svc): State<Arc<ServiceState>>,
-    Path(node_id): Path<String>,
-) -> impl IntoResponse {
-    match svc.app.store.with_read(|store| store.load_node_history(&node_id)) {
-        Ok(history) => Json(json!({ "node_id": node_id, "history": history })).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to load history" }))).into_response(),
-    }
-}
-
-async fn get_node_flap_status(
-    State(svc): State<Arc<ServiceState>>,
-    Path(node_id): Path<String>,
-) -> impl IntoResponse {
-    let five_minutes_ago = now_ms().saturating_sub(300_000);
-    match svc.app.store.with_read(|store| store.count_recent_posture_events(&node_id, five_minutes_ago)) {
-        Ok(count) => {
-            let status = FlapStatus {
-                node_id: node_id.clone(),
-                flapping: count >= 3,
-                event_count_5m: count,
-            };
-            Json(status).into_response()
-        }
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to query events" }))).into_response(),
-    }
-}
-
-async fn verify_audit_chain(
-    State(svc): State<Arc<ServiceState>>,
-) -> impl IntoResponse {
-    // `VerifyingKey` is `Copy`; copy it into the blocking task. A full-chain scan
-    // with a per-row Ed25519 verification is the heaviest read-side op — `call_read`
-    // runs it off the worker pool AND on a read-only replica, so it neither pins a
-    // tokio worker nor contends the writer mutex.
-    let vk = svc.audit_verifying_key;
-    let result = svc.app.store.call_read(move |store| {
-        store.verify_audit_chain_full(vk.as_ref())
-    })
-    .await;
-    match result {
-        Ok(Ok(r)) => Json(json!({
-            "chain_intact": r.chain_intact,
-            "total_entries": r.total_entries,
-            "latest_hash": r.latest_hash,
-            "signing_enabled": r.signing_enabled,
-            "signed_entries": r.signed_entries,
-            "unsigned_entries": r.unsigned_entries,
-            "signature_valid": r.signature_valid,
-            "first_signed_at_ms": r.first_signed_at_ms,
-            "public_key_b64": r.public_key_b64,
-            // #77 anchor-head high-water mark: detects tail truncation/deletion.
-            "head_verified": r.head_verified,
-            "head_status": r.head_status,
-            // Overall verdict folds in the head check so a truncated chain
-            // (rows internally consistent but tail deleted) reads as not-verified.
-            "verified": r.chain_intact && r.signature_valid && r.head_verified,
-        })).into_response(),
-        Ok(Err(_)) => (StatusCode::INTERNAL_SERVER_ERROR,
-                       Json(json!({ "error": "audit chain query failed" }))).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "store lock poisoned" }))).into_response(),
-    }
-}
 
 #[derive(Deserialize)]
 struct AuditExportQuery {
@@ -999,19 +579,6 @@ struct AuditExportQuery {
     offset: Option<u64>,
 }
 
-async fn handle_audit_export(
-    State(svc): State<Arc<ServiceState>>,
-    Query(params): Query<AuditExportQuery>,
-) -> impl IntoResponse {
-    let limit = params.limit.unwrap_or(100).min(1000);
-    let offset = params.offset.unwrap_or(0);
-    let vk = svc.audit_verifying_key.as_ref();
-    match svc.app.store.with_read(|store| store.load_audit_chain_page(limit, offset, vk)) {
-        Ok(page) => Json(page).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "export query failed" }))).into_response(),
-    }
-}
 
 #[derive(Deserialize)]
 struct RotateSigningKeyRequest {
@@ -1023,141 +590,7 @@ struct RotateSigningKeyRequest {
     reason: String,
 }
 
-async fn handle_audit_rotate_key(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<RotateSigningKeyRequest>,
-) -> impl IntoResponse {
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    // Decode the new signing seed → SigningKey (32-byte Ed25519 seed).
-    let new_signing_key = {
-        use base64::{engine::general_purpose::STANDARD as b64e, Engine as _};
-        match b64e.decode(req.new_signing_key_b64.trim())
-            .ok()
-            .and_then(|b| <[u8; 32]>::try_from(b.as_slice()).ok())
-            .map(|seed| ed25519_dalek::SigningKey::from_bytes(&seed))
-        {
-            Some(sk) => sk,
-            None => return (StatusCode::BAD_REQUEST,
-                Json(json!({ "error": "new_signing_key_b64 must be a base64 32-byte ed25519 seed" }))).into_response(),
-        }
-    };
-    let new_key_id = kirra_verifier::audit_chain::verifying_key_id(&new_signing_key.verifying_key());
-    // #79: pass our held fencing token so the durable write re-checks it INSIDE
-    // the transaction, closing the gate→commit TOCTOU.
-    let held_epoch = svc.app.held_epoch.load(std::sync::atomic::Ordering::SeqCst);
-    // The mutation runs under one acquisition; the closure returns the
-    // record_key_rotation result. The Fenced self-demote (a mode_active store)
-    // happens OUTSIDE the closure — the lock is already released by then,
-    // matching the prior `drop(store)`-before-self-demote ordering (Rule 4).
-    let rotation = svc.app.store.with(|store| {
-        store.record_key_rotation(new_signing_key, &req.reason, now_ms(), held_epoch)
-    });
-    match rotation {
-        Ok(_) => Json(json!({ "recorded": true, "event_type": "KEY_ROTATION", "new_key_id": new_key_id })).into_response(),
-        Err(DurableWriteError::Fenced(reason)) => {
-            // Superseded between the request-path gate and this commit.
-            // Mirror the gate: self-demote and reject fail-closed (no write
-            // landed). Subsequent mutations hit the standby check above.
-            svc.app.mode_active.store(false, std::sync::atomic::Ordering::SeqCst);
-            tracing::error!(
-                path = "/system/audit/rotate-signing-key",
-                fence = ?reason,
-                "FENCED at top-tier write (in-transaction epoch re-check) — self-demoting to PassiveStandby and rejecting"
-            );
-            (StatusCode::SERVICE_UNAVAILABLE,
-             Json(json!({ "error": "fenced: epoch superseded; instance demoted to passive standby" }))).into_response()
-        }
-        // NonceReplay cannot arise here (key rotation never touches the nonce
-        // table); fold it into the generic server-error arm for exhaustiveness.
-        Err(DurableWriteError::Db(_) | DurableWriteError::NonceReplay) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to record key rotation" }))).into_response(),
-    }
-}
 
-async fn evaluate_action_filter(
-    State(svc): State<Arc<ServiceState>>,
-    body: Result<Json<ActionClaim>, JsonRejection>,
-) -> impl IntoResponse {
-    let claim = match body {
-        Ok(Json(c)) => c,
-        Err(rejection) => {
-            // P1: durable audit write off the worker pool. Materialize the payload
-            // first (rejection is reused for the response detail below).
-            let now = now_ms();
-            let payload = json!({ "error": rejection.body_text() }).to_string();
-            let _ = svc.app.store.call(move |store| {
-                let _ = store.save_posture_event_chained(
-                    "action_filter", "ACTION_FILTER_MALFORMED_REQUEST",
-                    &payload,
-                    Some("malformed request body"), now,
-                );
-            }).await;
-            return (StatusCode::BAD_REQUEST, Json(json!({
-                "error": "MALFORMED_REQUEST",
-                "detail": rejection.body_text(),
-                "allowed": false,
-            }))).into_response();
-        }
-    };
-
-    let request_id = now_ms().to_string();
-
-    let (posture, lockout_reason) = gate_posture(&svc);
-
-    let posture_str = format!("{:?}", posture);
-    let decision = evaluate_action_claim(claim.clone(), posture);
-
-    let audit_event_type = if !decision.allowed {
-        if decision.reason == "UNKNOWN_ACTION_TYPE" {
-            "ACTION_FILTER_UNKNOWN_TYPE"
-        } else {
-            "ACTION_FILTER_DENIED"
-        }
-    } else {
-        "ACTION_FILTER_ALLOWED"
-    };
-
-    let event = json!({
-        "request_id": request_id,
-        "target_node": claim.target_node,
-        "action_type": claim.action_type,
-        "risk_class": claim.risk_class,
-        "allowed": decision.allowed,
-        "reason": decision.reason,
-        "posture": posture_str,
-        "lockout_reason": lockout_reason.as_ref().map(|r| r.to_string()),
-    });
-    // P1: durable audit write off the worker pool. `audit_event_type` is
-    // `&'static str` and the event string is owned, so both move into the closure.
-    let now = now_ms();
-    let event_str = event.to_string();
-    let _ = svc.app.store.call(move |store| {
-        let _ = store.save_posture_event_chained(
-            "action_filter", audit_event_type,
-            &event_str, None, now,
-        );
-    }).await;
-
-    tracing::info!(
-        action_type = %claim.action_type,
-        target_node = %claim.target_node,
-        allowed = decision.allowed,
-        posture = %posture_str,
-        reason = %decision.reason,
-        request_id = %request_id,
-        "action_filter evaluated"
-    );
-
-    Json(json!({
-        "allowed": decision.allowed,
-        "reason": decision.reason,
-        "posture_at_evaluation": posture_str,
-        "request_id": request_id,
-    })).into_response()
-}
 
 /// Replay/freshness metadata required on every per-protocol industrial request.
 /// Flattened on the wire: `sequence` and `timestamp_ms` sit at the top level
@@ -1172,547 +605,12 @@ struct ReplayGuarded<T> {
     message: T,
 }
 
-/// Enforce industrial-message replay + freshness (IEC 62443). Returns `Some(reason)`
-/// to REJECT (fail-closed), or `None` to proceed — advancing the per-source sequence
-/// high-water mark atomically on accept. Freshness is checked BEFORE the sequence
-/// advance, so a stale/future message never burns sequence space. A rejection is
-/// audit-chained. `source_id` should be caller-namespaced (e.g. `"dnp3:plc-01"`) so
-/// the same source string under different protocols does not collide.
-async fn enforce_industrial_replay(
-    svc: &ServiceState,
-    protocol: &str,
-    source_id: &str,
-    sequence: u64,
-    timestamp_ms: u64,
-) -> Option<&'static str> {
-    let now = now_ms();
-    let fresh = kirra_verifier::protocol_adapter::classify_industrial_freshness(
-        timestamp_ms,
-        now,
-        kirra_verifier::protocol_adapter::INDUSTRIAL_FRESHNESS_WINDOW_MS,
-    );
-    // The seq check-and-advance and the rejection audit write share ONE store
-    // acquisition (Rule 5: keep the read-then-write atomic). P1: run that whole
-    // group OFF the worker pool (`call` → spawn_blocking) — it fires on EVERY
-    // industrial command (the seq advance is itself a write), so it's the hottest
-    // durable path among the SCADA handlers. The single-acquisition atomicity is
-    // preserved (the entire closure runs under one writer lock on the blocking
-    // thread). Own the borrowed protocol/source strings into the closure; a task
-    // failure → fail-closed reject (store unavailable).
-    let protocol_owned = protocol.to_string();
-    let source_owned = source_id.to_string();
-    match svc.app.store.call(move |store| {
-        let reason = match fresh {
-            Some(r) => Some(r),
-            None => match store.industrial_seq_check_and_advance(&source_owned, sequence, now) {
-                Ok(true) => None,
-                Ok(false) => Some("INDUSTRIAL_MESSAGE_REPLAY"),
-                Err(_) => Some("INDUSTRIAL_REPLAY_STORE_UNAVAILABLE"),
-            },
-        };
-        if let Some(r) = reason {
-            let payload = json!({
-                "protocol": protocol_owned, "source_id": source_owned,
-                "sequence": sequence, "timestamp_ms": timestamp_ms, "reason": r,
-            });
-            let _ = store.save_posture_event_chained(
-                "industrial_replay_guard", "INDUSTRIAL_MESSAGE_REJECTED",
-                &payload.to_string(), Some(r), now,
-            );
-        }
-        reason
-    }).await {
-        Ok(reason) => reason,
-        Err(_) => Some("INDUSTRIAL_REPLAY_STORE_UNAVAILABLE"),
-    }
-}
 
-/// Standard rejection response for a replay/freshness denial (200 + allowed:false,
-/// matching the industrial handlers' denial shape; the rejection precedes evaluation).
-fn industrial_replay_rejection(protocol: &str, reason: &str) -> axum::response::Response {
-    (StatusCode::OK, Json(json!({
-        "protocol": protocol,
-        "allowed": false,
-        "denial_reason": reason,
-        "replay_rejected": true,
-    }))).into_response()
-}
 
-async fn evaluate_industrial_adapter(
-    State(svc): State<Arc<ServiceState>>,
-    body: Result<Json<UnifiedIndustrialRequest>, axum::extract::rejection::JsonRejection>,
-) -> impl IntoResponse {
-    let req = match body {
-        Ok(Json(r)) => r,
-        Err(rejection) => {
-            return (StatusCode::BAD_REQUEST, Json(json!({
-                "error": "MALFORMED_REQUEST",
-                "detail": rejection.body_text(),
-                "allowed": false,
-            }))).into_response();
-        }
-    };
 
-    let protocol_name = format!("{:?}", req.protocol);
 
-    // Replay/freshness gate (IEC 62443) — reject a stale/replayed message before
-    // evaluation. Key the per-source sequence by protocol:source_id.
-    let replay_key = format!("{protocol_name}:{}", req.source_id);
-    if let Some(reason) = enforce_industrial_replay(&svc, &protocol_name, &replay_key, req.sequence, req.timestamp_ms).await {
-        return industrial_replay_rejection(&protocol_name, reason);
-    }
 
-    let (posture, lockout_reason) = gate_posture(&svc);
 
-    let audit_ref = now_ms().to_string();
-
-    match evaluate_unified_industrial_request(req, posture) {
-        Ok(mut result) => {
-            // SG-012 / H-011: a DNP3 broadcast (only DNP3 sets `is_broadcast`)
-            // must carry a tamper-evident record; mirror the dedicated DNP3
-            // handler's fail-closed policy on this generic path too.
-            let is_broadcast = result.adapter_details
-                .get("is_broadcast").and_then(|v| v.as_bool()).unwrap_or(false);
-            let should_audit = !result.allowed || is_broadcast;
-
-            if should_audit {
-                let audit = json!({
-                    "protocol": result.protocol,
-                    "command": format!("{:?}", result.command),
-                    "allowed": result.allowed,
-                    "denial_reason": result.denial_reason,
-                    "posture": result.posture_at_evaluation,
-                    "lockout_reason": lockout_reason.as_ref().map(|r| r.to_string()),
-                    "audit_ref": audit_ref,
-                });
-                let event_type = if result.allowed {
-                    "INDUSTRIAL_ACTION_ALLOWED_BROADCAST"
-                } else {
-                    "INDUSTRIAL_ACTION_DENIED"
-                };
-                // P1: durable audit-chain write off the worker pool. Materialize
-                // the payload string first (move-able); a task failure → `false`
-                // (fail-closed), identical to the DB-error arm.
-                let now = now_ms();
-                let audit_str = audit.to_string();
-                let audit_ok = svc.app.store.call(move |store| match store.save_posture_event_chained(
-                    "industrial_adapter", event_type,
-                    &audit_str, None, now,
-                ) {
-                    Ok(()) => true,
-                    Err(e) => {
-                        tracing::error!(error = %e, event_type = event_type,
-                            "AUDIT-CHAIN WRITE FAILED for industrial adapter event — event missing from tamper-evident log");
-                        false
-                    }
-                }).await.unwrap_or_else(|e| {
-                    tracing::error!(error = %e, event_type = event_type,
-                        "AUDIT-CHAIN WRITE TASK FAILED for industrial adapter event — treating as unwritten (fail-closed)");
-                    false
-                });
-
-                // TR-012a: a broadcast whose mandatory audit could not be
-                // written is BLOCKED (fail-closed); non-broadcast audit failure
-                // stays non-fatal (TR-012b).
-                if is_broadcast && !audit_ok {
-                    result.allowed = false;
-                    result.denial_reason = Some("DNP3_BROADCAST_AUDIT_UNAVAILABLE".to_string());
-                    tracing::error!(
-                        "DNP3 broadcast BLOCKED (unified path) — mandatory audit write unavailable (SG-012 / H-011 fail-closed)");
-                }
-            }
-
-            Json(json!({
-                "protocol": result.protocol,
-                "command": format!("{:?}", result.command),
-                "allowed": result.allowed,
-                "denial_reason": result.denial_reason,
-                "posture_at_evaluation": result.posture_at_evaluation,
-                "adapter_details": result.adapter_details,
-                "audit_ref": audit_ref,
-                "triggers_recalculation": result.triggers_recalculation,
-            })).into_response()
-        }
-        Err(e) => {
-            (StatusCode::BAD_REQUEST, Json(json!({
-                "error": "ADAPTER_PARSE_FAILURE",
-                "detail": e,
-                "protocol": protocol_name,
-                "allowed": false,
-            }))).into_response()
-        }
-    }
-}
-
-async fn evaluate_ethernet_ip_adapter(
-    State(svc): State<Arc<ServiceState>>,
-    body: Result<Json<ReplayGuarded<EtherNetIpMessage>>, axum::extract::rejection::JsonRejection>,
-) -> impl IntoResponse {
-    let (msg, sequence, timestamp_ms) = match body {
-        Ok(Json(g)) => (g.message, g.sequence, g.timestamp_ms),
-        Err(rejection) => {
-            return (StatusCode::BAD_REQUEST, Json(json!({
-                "error": "MALFORMED_REQUEST",
-                "detail": rejection.body_text(),
-                "allowed": false,
-            }))).into_response();
-        }
-    };
-
-    // Replay/freshness gate before evaluation.
-    let replay_key = format!("ethernet_ip:{}", msg.source_node);
-    if let Some(reason) = enforce_industrial_replay(&svc, "ethernet_ip", &replay_key, sequence, timestamp_ms).await {
-        return industrial_replay_rejection("ethernet_ip", reason);
-    }
-
-    let (posture, lockout_reason) = gate_posture(&svc);
-
-    let posture_str = format!("{:?}", posture);
-    let eval = EtherNetIpAdapter::evaluate(&msg);
-    let (allowed, denial_reason) = kirra_verifier::protocol_adapter::command_allowed_for_posture_pub(&eval.command, &posture);
-    let audit_ref = now_ms().to_string();
-
-    if !allowed {
-        // P1: durable audit write off the worker pool. Materialize the payload
-        // string first so the closure owns it (`eval`/`denial_reason` are reused
-        // in the response below).
-        let now = now_ms();
-        let payload = json!({
-            "service_name": eval.service_name,
-            "safety_relevant": eval.safety_relevant,
-            "posture": posture_str,
-            "denial_reason": denial_reason,
-            "lockout_reason": lockout_reason.as_ref().map(|r| r.to_string()),
-        }).to_string();
-        let _ = svc.app.store.call(move |store| {
-            if let Err(e) = store.save_posture_event_chained(
-                "ethernet_ip_adapter", "INDUSTRIAL_ACTION_DENIED",
-                &payload, None, now,
-            ) {
-                tracing::error!(error = %e, event_type = "INDUSTRIAL_ACTION_DENIED",
-                    "AUDIT-CHAIN WRITE FAILED for ethernet_ip adapter event — event missing from tamper-evident log");
-            }
-        }).await;
-    }
-
-    Json(json!({
-        "protocol": "ethernet_ip",
-        "command": format!("{:?}", eval.command),
-        "allowed": allowed,
-        "denial_reason": denial_reason,
-        "posture_at_evaluation": posture_str,
-        "adapter_details": {
-            "service_name": eval.service_name,
-            "is_write": eval.is_write,
-            "target_description": eval.target_description,
-            "safety_relevant": eval.safety_relevant,
-        },
-        "audit_ref": audit_ref,
-    })).into_response()
-}
-
-async fn evaluate_canopen_adapter(
-    State(svc): State<Arc<ServiceState>>,
-    body: Result<Json<ReplayGuarded<CanOpenMessage>>, axum::extract::rejection::JsonRejection>,
-) -> impl IntoResponse {
-    let (msg, sequence, timestamp_ms) = match body {
-        Ok(Json(g)) => (g.message, g.sequence, g.timestamp_ms),
-        Err(rejection) => {
-            return (StatusCode::BAD_REQUEST, Json(json!({
-                "error": "MALFORMED_REQUEST",
-                "detail": rejection.body_text(),
-                "allowed": false,
-            }))).into_response();
-        }
-    };
-
-    // Replay/freshness gate before evaluation.
-    let replay_key = format!("canopen:{}", msg.source_node);
-    if let Some(reason) = enforce_industrial_replay(&svc, "canopen", &replay_key, sequence, timestamp_ms).await {
-        return industrial_replay_rejection("canopen", reason);
-    }
-
-    let (posture, lockout_reason) = gate_posture(&svc);
-
-    let posture_str = format!("{:?}", posture);
-    let eval = CanOpenAdapter::evaluate(&msg);
-    let (allowed, denial_reason) = kirra_verifier::protocol_adapter::command_allowed_for_posture_pub(&eval.command, &posture);
-    let audit_ref = now_ms().to_string();
-
-    // #84: resolve the CANopen bus node-id to a FLEET node so an NMT-offline
-    // event marks the correct asset and the recalc is EFFECTFUL. Unmapped or
-    // unregistered ids are FAIL-CLOSED — surfaced as an unattributed offline
-    // (distinct audit event + warning + response flag), never a silent no-op.
-    let offline_outcome = if eval.triggers_recalculation {
-        use kirra_verifier::adapters::canopen::{classify_nmt_offline, global_resolve};
-        let resolved = global_resolve(eval.node_id);
-        let registered = resolved
-            .as_deref()
-            .map(|n| svc.app.nodes.contains_key(n))
-            .unwrap_or(false);
-        Some(classify_nmt_offline(eval.node_id, resolved, registered))
-    } else {
-        None
-    };
-
-    // Apply the offline effect (mark the node + drive a recalc). The fleet node
-    // actually marked offline (if any) is recorded for the audit + response.
-    let mut attributed_fleet_node: Option<String> = None;
-    if let Some(outcome) = &offline_outcome {
-        use kirra_verifier::adapters::canopen::{NmtOfflineOutcome, UnattributedReason};
-        use kirra_verifier::posture_engine_v2::PostureRecalcTrigger;
-        match outcome {
-            NmtOfflineOutcome::Attributed { fleet_node_id } => {
-                match svc.app.mark_node_untrusted(fleet_node_id, "CANOPEN_NMT_OFFLINE", now_ms()) {
-                    Ok(true) => {
-                        attributed_fleet_node = Some(fleet_node_id.clone());
-                        tracing::warn!(
-                            canopen_node_id = eval.node_id,
-                            fleet_node_id = %fleet_node_id,
-                            "CANopen NMT node-offline → fleet node marked Untrusted; effectful recalc enqueued"
-                        );
-                        enqueue_recalc(&svc, PostureRecalcTrigger::NodeTrustChanged {
-                            node_id: fleet_node_id.clone(),
-                            reason: "CANOPEN_NMT_OFFLINE".to_string(),
-                        });
-                    }
-                    // Mapping raced a deregistration, or the store write failed:
-                    // fail-closed exactly like an unattributed offline.
-                    Ok(false) | Err(()) => {
-                        tracing::error!(
-                            canopen_node_id = eval.node_id,
-                            fleet_node_id = %fleet_node_id,
-                            "CANopen NMT node-offline: mapped node missing or store write failed — \
-                             treating as UNATTRIBUTED (fail-closed)"
-                        );
-                        enqueue_recalc(&svc, PostureRecalcTrigger::DependencyGraphChanged);
-                    }
-                }
-            }
-            NmtOfflineOutcome::Unattributed { canopen_node_id, reason } => {
-                let reason_str = match reason {
-                    UnattributedReason::NoMapping => "NO_MAPPING",
-                    UnattributedReason::NodeNotRegistered => "NODE_NOT_REGISTERED",
-                };
-                tracing::warn!(
-                    canopen_node_id = *canopen_node_id,
-                    source_node = %msg.source_node,
-                    reason = reason_str,
-                    "CANopen NMT node-offline UNATTRIBUTED — recorded as unattributed offline; \
-                     recalc enqueued (fail-closed, never silently dropped)"
-                );
-                enqueue_recalc(&svc, PostureRecalcTrigger::DependencyGraphChanged);
-            }
-        }
-    }
-
-    if !allowed || eval.triggers_recalculation {
-        // P1: durable audit write off the worker pool. `event_type` (&'static str)
-        // and the payload string are materialized first so the closure owns them
-        // (`eval` is reused in the response below).
-        let now = now_ms();
-        let event_type = if eval.triggers_recalculation {
-            if attributed_fleet_node.is_some() {
-                "CANOPEN_NMT_NODE_OFFLINE"
-            } else {
-                "CANOPEN_NMT_OFFLINE_UNATTRIBUTED"
-            }
-        } else {
-            "INDUSTRIAL_ACTION_DENIED"
-        };
-        let payload = json!({
-            "node_id": eval.node_id,
-            "fleet_node_id": attributed_fleet_node.clone(),
-            "node_offline_attributed": attributed_fleet_node.is_some(),
-            "message_type": format!("{:?}", eval.message_type),
-            "is_emergency": eval.is_emergency,
-            "triggers_recalculation": eval.triggers_recalculation,
-            "posture": posture_str,
-            "lockout_reason": lockout_reason.as_ref().map(|r| r.to_string()),
-        }).to_string();
-        let _ = svc.app.store.call(move |store| {
-            if let Err(e) = store.save_posture_event_chained(
-                "canopen_adapter", event_type,
-                &payload, None, now,
-            ) {
-                tracing::error!(error = %e, event_type = event_type,
-                    "AUDIT-CHAIN WRITE FAILED for canopen adapter event — event missing from tamper-evident log");
-            }
-        }).await;
-    }
-
-    Json(json!({
-        "protocol": "canopen",
-        "command": format!("{:?}", eval.command),
-        "allowed": allowed,
-        "denial_reason": denial_reason,
-        "posture_at_evaluation": posture_str,
-        // #84: whether the NMT-offline was attributed to a fleet node (and which).
-        // `false` on an offline event means the id was unmapped/unregistered and
-        // handled fail-closed — surfaced here so callers never read silence as success.
-        "node_offline_attributed": attributed_fleet_node.is_some(),
-        "fleet_node_id": attributed_fleet_node,
-        "adapter_details": {
-            "message_type": format!("{:?}", eval.message_type),
-            "node_id": eval.node_id,
-            "is_emergency": eval.is_emergency,
-            "emergency_code": eval.emergency_code,
-            "triggers_recalculation": eval.triggers_recalculation,
-        },
-        "audit_ref": audit_ref,
-    })).into_response()
-}
-
-async fn evaluate_dnp3_adapter(
-    State(svc): State<Arc<ServiceState>>,
-    body: Result<Json<ReplayGuarded<Dnp3Message>>, axum::extract::rejection::JsonRejection>,
-) -> impl IntoResponse {
-    let (msg, sequence, timestamp_ms) = match body {
-        Ok(Json(g)) => (g.message, g.sequence, g.timestamp_ms),
-        Err(rejection) => {
-            return (StatusCode::BAD_REQUEST, Json(json!({
-                "error": "MALFORMED_REQUEST",
-                "detail": rejection.body_text(),
-                "allowed": false,
-            }))).into_response();
-        }
-    };
-
-    // Replay/freshness gate before evaluation.
-    let replay_key = format!("dnp3:{}", msg.source_node);
-    if let Some(reason) = enforce_industrial_replay(&svc, "dnp3", &replay_key, sequence, timestamp_ms).await {
-        return industrial_replay_rejection("dnp3", reason);
-    }
-
-    let (posture, lockout_reason) = gate_posture(&svc);
-
-    let posture_str = format!("{:?}", posture);
-    let eval = Dnp3Adapter::evaluate(&msg);
-    let (allowed, denial_reason) = kirra_verifier::protocol_adapter::command_allowed_for_posture_pub(&eval.command, &posture);
-    let audit_ref = now_ms().to_string();
-
-    // SG-012 / H-011 — a DNP3 broadcast control command must carry a
-    // tamper-evident record. Kirra CLASSIFIES, it does not actuate: "before
-    // control output" means before this handler returns its verdict, and the
-    // integrator MUST NOT actuate ahead of this audited verdict (a documented
-    // assumption-of-use). Audit policy:
-    //   * Broadcast (TR-012 / TR-012a): MUST be audited; if the mandatory audit
-    //     write fails (or the store lock is poisoned), the command is BLOCKED
-    //     (fail-closed) — H-011's hazard is a broadcast control executed without
-    //     a tamper-evident record.
-    //   * Unicast control (TR-012b): also audited, but an audit-write failure is
-    //     NON-fatal — for a single target the enforcement decision outranks the
-    //     record (blocking on a transient disk error would be fail-open).
-    //   * Denials: audited.
-    let mut allowed = allowed;
-    let mut denial_reason = denial_reason;
-    let mut status = StatusCode::OK;
-
-    // MAGNITUDE BOUND: a posture-admitted Analog Output (g41) control must also lie
-    // within the configured envelope — else REFUSE (fail-closed: also when no
-    // envelope is configured or the value is undecodable). Only applied on the
-    // posture-allowed path; a posture denial outranks and is reported as-is. The
-    // override-to-denied flows into the audit block below as an action-denied event.
-    if allowed {
-        if let Err(reason) = Dnp3Adapter::bound_analog_control(
-            &msg,
-            kirra_verifier::adapters::dnp3::global_analog_envelope().as_ref(),
-        ) {
-            allowed = false;
-            denial_reason = Some(reason.to_string());
-        }
-    }
-
-    if eval.is_broadcast || eval.is_control || !allowed {
-        let event_type = if eval.is_broadcast {
-            "DNP3_BROADCAST_COMMAND"
-        } else if !allowed {
-            "INDUSTRIAL_ACTION_DENIED"
-        } else {
-            "DNP3_CONTROL_COMMAND"
-        };
-        let audit_payload = json!({
-            "function_name": eval.function_name,
-            "is_broadcast": eval.is_broadcast,
-            "is_control": eval.is_control,
-            "critical_infrastructure_relevant": eval.critical_infrastructure_relevant,
-            "dest_address": msg.dest_address,
-            "posture": posture_str,
-            "lockout_reason": lockout_reason.as_ref().map(|r| r.to_string()),
-        })
-        .to_string();
-        // P1: the per-command audit-chain write (durable fsync) runs off the
-        // worker pool. `event_type` is `&'static str` and the payload is owned, so
-        // both move into the closure cleanly. A task failure → `false` (fail-closed:
-        // audit unconfirmed), identical to the DB-error arm below.
-        let now = now_ms();
-        let audit_payload_owned = audit_payload;
-        let audit_ok = svc.app.store.call(move |store| match store.save_posture_event_chained(
-            "dnp3_adapter", event_type, &audit_payload_owned, None, now,
-        ) {
-            Ok(()) => true,
-            Err(e) => {
-                tracing::error!(error = %e, event_type = event_type,
-                    "AUDIT-CHAIN WRITE FAILED for dnp3 adapter event — event missing from tamper-evident log");
-                false
-            }
-        }).await.unwrap_or_else(|e| {
-            tracing::error!(error = %e, event_type = event_type,
-                "AUDIT-CHAIN WRITE TASK FAILED for dnp3 adapter event — treating as unwritten (fail-closed)");
-            false
-        });
-
-        // TR-012a: a BROADCAST whose mandatory audit could not be written is
-        // BLOCKED (fail-closed). Unicast audit failure is non-fatal (TR-012b).
-        if eval.is_broadcast && !audit_ok {
-            allowed = false;
-            denial_reason = Some("DNP3_BROADCAST_AUDIT_UNAVAILABLE".to_string());
-            status = StatusCode::SERVICE_UNAVAILABLE;
-            tracing::error!(dest_address = msg.dest_address,
-                "DNP3 BROADCAST control BLOCKED — mandatory audit write unavailable (SG-012 / H-011 fail-closed)");
-        }
-    }
-
-    (status, Json(json!({
-        "protocol": "dnp3",
-        "command": format!("{:?}", eval.command),
-        "allowed": allowed,
-        "denial_reason": denial_reason,
-        "posture_at_evaluation": posture_str,
-        "adapter_details": {
-            "function_name": eval.function_name,
-            "is_control": eval.is_control,
-            "is_broadcast": eval.is_broadcast,
-            "critical_infrastructure_relevant": eval.critical_infrastructure_relevant,
-        },
-        "audit_ref": audit_ref,
-    }))).into_response()
-}
-
-async fn register_federation_controller(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<RegisterFederationControllerRequest>,
-) -> impl IntoResponse {
-    if req.controller_id.trim().is_empty() || req.public_key_b64.trim().is_empty() {
-        return (StatusCode::BAD_REQUEST,
-                Json(json!({ "error": "controller_id and public_key_b64 are required" }))).into_response();
-    }
-    // P1: durable write off the worker pool (`call` → spawn_blocking). Own the
-    // captured request fields (still needed for the response). `call` wraps the
-    // closure's own `Result` in an outer `Result<_, StoreError>`, so a DB error
-    // OR a task failure both map to 500 (fail-closed, unchanged from `with`).
-    let now = now_ms();
-    let controller_id = req.controller_id.clone();
-    let public_key_b64 = req.public_key_b64.clone();
-    match svc.app.store.call(move |store| store.save_trusted_federation_controller(
-        &controller_id, &public_key_b64, now,
-    )).await {
-        Ok(Ok(())) => (StatusCode::CREATED,
-                   Json(json!({ "controller_id": req.controller_id, "registered": true }))).into_response(),
-        _ => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to register controller" }))).into_response(),
-    }
-}
 
 #[derive(Deserialize)]
 struct RegisterIdentityRequest {
@@ -1720,445 +618,11 @@ struct RegisterIdentityRequest {
     ak_public_fingerprint_hex: String,
 }
 
-async fn register_node_identity(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<RegisterIdentityRequest>,
-) -> impl IntoResponse {
-    if req.node_id.trim().is_empty() || req.ak_public_fingerprint_hex.trim().is_empty() {
-        return (StatusCode::BAD_REQUEST,
-                Json(json!({ "error": "node_id and ak_public_fingerprint_hex are required" }))).into_response();
-    }
-    // P1: durable identity write off the worker pool. Own the request fields.
-    let now = now_ms();
-    let node_id_c = req.node_id.clone();
-    let fingerprint = req.ak_public_fingerprint_hex.clone();
-    let registered = svc.app.store.call(move |store| store.register_attestation_identity(
-        &node_id_c, &fingerprint, "admin", now,
-    )).await;
-    match registered {
-        Ok(Ok(())) => {
-            emit_posture_event(&svc.app, "NODE_IDENTITY_PROVISIONED", Some(req.node_id.clone()));
-            (StatusCode::CREATED,
-             Json(json!({ "node_id": req.node_id, "registered": true }))).into_response()
-        }
-        _ => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "failed to register identity" }))).into_response(),
-    }
-}
 
-async fn submit_federated_report(
-    State(svc): State<Arc<ServiceState>>,
-    // v2 wire: `source_generation` is optional, so a v1 report (no such field)
-    // deserializes to a V2 with `source_generation: None` and its signed payload
-    // is byte-identical to the v1 canonical payload — backward compatible. The
-    // generation, when present, is inside the signed payload (cannot be forged
-    // or stripped) and drives generation-ordered conflict resolution at read time.
-    Json(report): Json<FederatedTrustReportV2>,
-) -> impl IntoResponse {
-    let received_at_ms = now_ms();
 
-    let evaluation = evaluate_federated_report_v2(&report, received_at_ms);
-    if !evaluation.accepted {
-        return Json(evaluation).into_response();
-    }
 
-    // #79: held fencing token, read before locking the store. The durable commit
-    // re-checks it INSIDE the transaction, closing the gate→commit TOCTOU.
-    let held_epoch = svc.app.held_epoch.load(std::sync::atomic::Ordering::SeqCst);
 
-    // The whole 5-step commit (key load → signature verify → nonce gate → chained
-    // report+nonce-burn commit) runs under ONE lock; offload it to the blocking pool
-    // so this multi-statement transaction — the heaviest write — can't pin a tokio
-    // worker. The #79 epoch fence is preserved: `held_epoch` is read above (before
-    // the lock, as before) and the durable commit re-checks it INSIDE its
-    // transaction, so the slightly larger read→lock window remains harmless. All
-    // rejection-path audit writes stay inside the locked closure (same atomicity).
-    let outcome = svc.app.store.call(move |store| {
-        let pk_b64 = match store.load_trusted_federation_controller_key(&report.source_controller_id) {
-            Ok(Some(key)) => key,
-            Ok(None) => {
-                let event = json!({ "source_controller_id": report.source_controller_id,
-                                    "reason": "UNREGISTERED_FEDERATION_CONTROLLER" });
-                let _ = store.save_posture_event_chained(
-                    "federation_gateway", "FEDERATION_REJECTED",
-                    &event.to_string(), Some("unregistered source"), received_at_ms,
-                );
-                return FedCommitOutcome::Rejected("UNREGISTERED_FEDERATION_CONTROLLER");
-            }
-            Err(_) => return FedCommitOutcome::InternalError("controller lookup failed"),
-        };
 
-        if !verify_federated_report_signature_v2(&report, &pk_b64) {
-            let event = json!({ "source_controller_id": report.source_controller_id,
-                                "reason": "INVALID_FEDERATION_SIGNATURE" });
-            let _ = store.save_posture_event_chained(
-                "federation_gateway", "FEDERATION_REJECTED",
-                &event.to_string(), Some("signature mismatch"), received_at_ms,
-            );
-            return FedCommitOutcome::Rejected("INVALID_FEDERATION_SIGNATURE");
-        }
-
-        match store.has_seen_federation_nonce(&report.nonce_hex) {
-            Ok(true) => {
-                let event = json!({ "source_controller_id": report.source_controller_id,
-                                    "nonce_hex": report.nonce_hex,
-                                    "reason": "FEDERATION_NONCE_REPLAY" });
-                let _ = store.save_posture_event_chained(
-                    "federation_gateway", "FEDERATION_REJECTED",
-                    &event.to_string(), Some("nonce replay"), received_at_ms,
-                );
-                return FedCommitOutcome::Rejected("FEDERATED_NONCE_REPLAY");
-            }
-            Ok(false) => {}
-            Err(_) => return FedCommitOutcome::InternalError("nonce lookup failed"),
-        }
-
-        match store.save_federated_report_chained(&report.as_v1(), report.source_generation, received_at_ms, held_epoch) {
-            Ok(()) => FedCommitOutcome::Accepted,
-            Err(DurableWriteError::NonceReplay) => {
-                // H1: a replay raced past the `has_seen_federation_nonce` gate above and
-                // lost the durable single-use claim (PRIMARY KEY violation aborted the
-                // transaction — report NOT persisted, nonce NOT double-burned). Map it to
-                // the SAME clean rejection + audit as the gate path, not an opaque 500.
-                let event = json!({ "source_controller_id": report.source_controller_id,
-                                    "nonce_hex": report.nonce_hex,
-                                    "reason": "FEDERATION_NONCE_REPLAY" });
-                let _ = store.save_posture_event_chained(
-                    "federation_gateway", "FEDERATION_REJECTED",
-                    &event.to_string(), Some("nonce replay (concurrent)"), received_at_ms,
-                );
-                FedCommitOutcome::Rejected("FEDERATED_NONCE_REPLAY")
-            }
-            Err(DurableWriteError::Fenced(reason)) => FedCommitOutcome::Fenced(format!("{reason:?}")),
-            Err(DurableWriteError::Db(_)) => FedCommitOutcome::InternalError("failed to persist federated report"),
-        }
-    })
-    .await;
-
-    let outcome = match outcome {
-        Ok(o) => o,
-        Err(_) => return (StatusCode::INTERNAL_SERVER_ERROR,
-                          Json(json!({ "error": "store lock poisoned" }))).into_response(),
-    };
-
-    match outcome {
-        FedCommitOutcome::Accepted => Json(evaluation).into_response(),
-        FedCommitOutcome::Rejected(reason) =>
-            Json(ReportEvaluation { accepted: false, reason: reason.to_string() }).into_response(),
-        FedCommitOutcome::InternalError(msg) =>
-            (StatusCode::INTERNAL_SERVER_ERROR, Json(json!({ "error": msg }))).into_response(),
-        FedCommitOutcome::Fenced(reason) => {
-            // Superseded between the request-path gate and this commit. Mirror the
-            // gate: self-demote and reject fail-closed — the report was NOT persisted
-            // and the nonce was NOT burned (the transaction was dropped in the closure).
-            svc.app.mode_active.store(false, std::sync::atomic::Ordering::SeqCst);
-            tracing::error!(
-                path = "/federation/reports/submit",
-                fence = %reason,
-                "FENCED at top-tier write (in-transaction epoch re-check) — self-demoting to PassiveStandby and rejecting"
-            );
-            (StatusCode::SERVICE_UNAVAILABLE,
-             Json(json!({ "error": "fenced: epoch superseded; instance demoted to passive standby" }))).into_response()
-        }
-    }
-}
-
-async fn get_federated_reports(
-    State(svc): State<Arc<ServiceState>>,
-    Path(asset_id): Path<String>,
-) -> impl IntoResponse {
-    // Both reads share ONE acquisition (Rule 5). The closure returns a Result so
-    // the per-read error responses are produced OUTSIDE the closure (Rule 4).
-    let loaded = svc.app.store.with_read(|store| {
-        let reports = store.load_federated_reports_for_asset(&asset_id)
-            .map_err(|_| "failed to load reports")?;
-        // #329 v2 — generation-ordered conflict resolution. Reconcile the stored
-        // reports into the single authoritative posture (higher generation wins;
-        // ties fall back to issued_at_ms, then fail closed to the more restrictive
-        // posture). `null` when no reports exist for the asset. This is a read-time
-        // view only — it does NOT feed the local posture engine that gates actuators.
-        let authoritative = store.load_federated_report_v2s_for_asset(&asset_id)
-            .map(|v2s| authoritative_posture(&v2s))
-            .map_err(|_| "failed to reconcile reports")?;
-        Ok::<_, &'static str>((reports, authoritative))
-    });
-    let (reports, authoritative) = match loaded {
-        Ok(pair) => pair,
-        Err(msg) => return (StatusCode::INTERNAL_SERVER_ERROR,
-                            Json(json!({ "error": msg }))).into_response(),
-    };
-
-    Json(json!({
-        "asset_id": asset_id,
-        "reports": reports,
-        "authoritative_posture": authoritative,
-    })).into_response()
-}
-
-async fn handle_actuator_motion_command(
-    State(svc): State<Arc<ServiceState>>,
-    // Threaded by `enforce_actuator_safety_envelope` (the route always runs it
-    // first). It carries the TRUE verdict — `cmd` below is already the enforced
-    // (post-clamp) command, but only this tells us WHETHER a clamp happened, so
-    // the response can report it instead of always claiming "Allow".
-    Extension(outcome): Extension<EnforcementOutcome>,
-    Json(cmd): Json<ProposedVehicleCommand>,
-) -> impl IntoResponse {
-    let now = now_ms();
-
-    tracing::info!(
-        action              = %outcome.action.as_str(),
-        linear_velocity_mps = %cmd.linear_velocity_mps,
-        steering_angle_deg  = %cmd.steering_angle_deg,
-        delta_time_s        = %cmd.delta_time_s,
-        "Actuator motion command admitted through safety envelope"
-    );
-
-    // Record the verdict in the tamper-evident log, including whether a clamp
-    // occurred and the original vs enforced values (previously a clamp was
-    // logged indistinguishably from a plain admit).
-    let audit = serde_json::json!({
-        "action":                       outcome.action.as_str(),
-        "original_linear_velocity_mps": outcome.original_linear_velocity_mps,
-        "original_steering_angle_deg":  outcome.original_steering_angle_deg,
-        "enforced_linear_velocity_mps": outcome.enforced_linear_velocity_mps,
-        "enforced_steering_angle_deg":  outcome.enforced_steering_angle_deg,
-        "current_velocity_mps":         cmd.current_velocity_mps,
-        "current_steering_angle_deg":   cmd.current_steering_angle_deg,
-        "delta_time_s":                 cmd.delta_time_s,
-        "admitted_at_ms":               now,
-    });
-    // P1: durable audit write off the worker pool (materialize the payload first).
-    let audit_str = audit.to_string();
-    let _ = svc.app.store.call(move |store| {
-        if let Err(e) = store.save_posture_event_chained(
-            "actuator_motion", "MOTION_COMMAND_ADMITTED",
-            &audit_str, None, now,
-        ) {
-            tracing::error!(error=%e,
-                "AUDIT-CHAIN WRITE FAILED for MOTION_COMMAND_ADMITTED — event missing from tamper-evident log");
-        }
-    }).await;
-
-    // Response speaks the ROS interceptor's schema (action / enforced_*) AND
-    // the legacy keys (now accurate). See `EnforcementOutcome::response_body`.
-    (StatusCode::OK, Json(outcome.response_body())).into_response()
-}
-
-async fn handle_sensor_fault_report(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<SensorFaultReportRequest>,
-) -> impl IntoResponse {
-    if req.source_node_id.trim().is_empty() {
-        return (StatusCode::BAD_REQUEST,
-                Json(json!({ "error": "source_node_id is required" }))).into_response();
-    }
-    if !svc.app.nodes.contains_key(&req.source_node_id) {
-        return (StatusCode::NOT_FOUND,
-                Json(json!({ "error": "node not registered" }))).into_response();
-    }
-
-    let now = now_ms();
-
-    let confidence_floor = svc.app.store.with(|store| {
-        store.load_av_confidence_floor(&req.source_node_id)
-            .unwrap_or(None)
-            .unwrap_or(0.70)
-    });
-
-    let is_degraded = req.hardware_fault_detected || req.confidence_score < confidence_floor;
-
-    if is_degraded {
-        let reason = if req.hardware_fault_detected { "hardware_fault" } else { "low_confidence" };
-
-        svc.app.store.with(|store| {
-            let _ = store.reset_recovery_streak(&req.source_node_id, now);
-        });
-
-        let updated = match svc.app.nodes.get(&req.source_node_id) {
-            Some(n) => RegisteredNode {
-                node_id:              n.node_id.clone(),
-                status:               NodeTrustState::Untrusted(reason.to_string()),
-                registered_at_ms:     n.registered_at_ms,
-                last_trust_update_ms: now,
-                ak_public_pem:        n.ak_public_pem.clone(),
-                expected_pcr16_digest_hex: n.expected_pcr16_digest_hex.clone(),
-                site:                 n.site.clone(),
-                firmware_version:     n.firmware_version.clone(),
-            },
-            None => return (StatusCode::NOT_FOUND,
-                            Json(json!({ "error": "node not found" }))).into_response(),
-        };
-
-        if svc.app.persist_and_insert_node(updated).is_err() {
-            return (StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(json!({ "error": "failed to persist node state" }))).into_response();
-        }
-
-        let event = json!({
-            "source_node_id":          req.source_node_id,
-            "confidence_score":        req.confidence_score,
-            "hardware_fault_detected": req.hardware_fault_detected,
-            "reason":                  reason,
-        });
-        // P1: durable audit write off the worker pool (own the node id + payload).
-        let event_str = event.to_string();
-        let node_id_c = req.source_node_id.clone();
-        let _ = svc.app.store.call(move |store| {
-            if let Err(e) = store.save_posture_event_chained(
-                &node_id_c, "SENSOR_HEALTH_REPORT_FAULT",
-                &event_str, None, now,
-            ) {
-                tracing::error!(error=%e, node_id=%node_id_c,
-                    "AUDIT-CHAIN WRITE FAILED for SENSOR_HEALTH_REPORT_FAULT — event missing from tamper-evident log");
-            }
-        }).await;
-
-        emit_posture_event(&svc.app, "NODE_STATUS_CHANGED", Some(req.source_node_id.clone()));
-        enqueue_recalc(&svc, kirra_verifier::posture_engine_v2::PostureRecalcTrigger::NodeTrustChanged {
-            node_id: req.source_node_id.clone(),
-            reason:  format!("SENSOR_FAULT:{reason}"),
-        });
-
-        return (StatusCode::OK, Json(json!({
-            "source_node_id": req.source_node_id,
-            "accepted": true,
-            "fault_recorded": true,
-        }))).into_response();
-    }
-
-    let currently_untrusted = svc.app.nodes.get(&req.source_node_id)
-        .map(|n| matches!(n.status, NodeTrustState::Untrusted(_)))
-        .unwrap_or(false);
-
-    if !currently_untrusted {
-        svc.app.store.with(|store| {
-            let _ = store.touch_av_telemetry_timestamp(&req.source_node_id, now);
-        });
-        return (StatusCode::OK, Json(json!({
-            "source_node_id": req.source_node_id,
-            "accepted": true,
-            "fault_recorded": false,
-        }))).into_response();
-    }
-
-    let decision = svc.app.store.with(|store| {
-        // `&*store` dereferences to `&VerifierStore` so the generic
-        // `S: RecoveryStreakStore` bound on `evaluate_recovery_report`
-        // resolves correctly (S3 / #115 — trait seam, behavior unchanged:
-        // the trait impl delegates verbatim).
-        evaluate_recovery_report(&*store, &req.source_node_id, now)
-    });
-
-    match &decision {
-        HysteresisDecision::RecoveryConfirmed { streak } => {
-            let updated = match svc.app.nodes.get(&req.source_node_id) {
-                Some(n) => RegisteredNode {
-                    node_id:              n.node_id.clone(),
-                    status:               NodeTrustState::Trusted,
-                    registered_at_ms:     n.registered_at_ms,
-                    last_trust_update_ms: now,
-                    ak_public_pem:        n.ak_public_pem.clone(),
-                    expected_pcr16_digest_hex: n.expected_pcr16_digest_hex.clone(),
-                    site:                 n.site.clone(),
-                    firmware_version:     n.firmware_version.clone(),
-                },
-                None => return (StatusCode::NOT_FOUND,
-                                Json(json!({ "error": "node not found" }))).into_response(),
-            };
-
-            if svc.app.persist_and_insert_node(updated).is_err() {
-                return (StatusCode::INTERNAL_SERVER_ERROR,
-                        Json(json!({ "error": "failed to persist node state" }))).into_response();
-            }
-
-            // P1: the recovery-streak reset + its audit write run as ONE off-worker
-            // closure (same single-acquisition grouping as before, just off the pool).
-            let node_id_c = req.source_node_id.clone();
-            let streak_v = *streak;
-            let _ = svc.app.store.call(move |store| {
-                let _ = store.reset_recovery_streak(&node_id_c, now);
-                let event = json!({
-                    "source_node_id": node_id_c.as_str(),
-                    "streak":         streak_v,
-                });
-                if let Err(e) = store.save_posture_event_chained(
-                    &node_id_c, "SENSOR_RECOVERY_CONFIRMED",
-                    &event.to_string(), None, now,
-                ) {
-                    tracing::error!(error=%e, node_id=%node_id_c,
-                        "AUDIT-CHAIN WRITE FAILED for SENSOR_RECOVERY_CONFIRMED — event missing from tamper-evident log");
-                }
-            }).await;
-
-            emit_posture_event(&svc.app, "NODE_STATUS_CHANGED", Some(req.source_node_id.clone()));
-            enqueue_recalc(&svc, kirra_verifier::posture_engine_v2::PostureRecalcTrigger::NodeTrustChanged {
-                node_id: req.source_node_id.clone(),
-                reason:  "SENSOR_RECOVERY_CONFIRMED".to_string(),
-            });
-        }
-        HysteresisDecision::StreakBuilding { .. } | HysteresisDecision::WindowExpired { .. } => {}
-        HysteresisDecision::NotApplicable => {
-            svc.app.store.with(|store| {
-                let _ = store.touch_av_telemetry_timestamp(&req.source_node_id, now);
-            });
-        }
-    }
-
-    (StatusCode::OK, Json(json!({
-        "source_node_id":      req.source_node_id,
-        "accepted":            true,
-        "fault_recorded":      false,
-        "hysteresis_decision": format!("{:?}", decision),
-    }))).into_response()
-}
-
-async fn handle_register_av_asset(
-    State(svc): State<Arc<ServiceState>>,
-    Json(req): Json<RegisterAvAssetRequest>,
-) -> impl IntoResponse {
-    if req.node_id.trim().is_empty() {
-        return (StatusCode::BAD_REQUEST,
-                Json(json!({ "error": "node_id is required" }))).into_response();
-    }
-    if !svc.app.nodes.contains_key(&req.node_id) {
-        return (StatusCode::NOT_FOUND,
-                Json(json!({ "error": "node not registered" }))).into_response();
-    }
-
-    let now = now_ms();
-    let floor = req.confidence_floor.unwrap_or(0.70);
-
-    // P1: registration + its audit write run as ONE off-worker closure (own the
-    // request fields; node id is reused in the response).
-    let node_id_c = req.node_id.clone();
-    let subsystem_type_c = req.subsystem_type.clone();
-    let hardware_id_c = req.hardware_id.clone();
-    let _ = svc.app.store.call(move |store| {
-        if let Err(e) = store.register_av_subsystem_meta(
-            &node_id_c, &subsystem_type_c, &hardware_id_c, floor, now,
-        ) {
-            tracing::warn!(
-                error   = %e,
-                node_id = %node_id_c,
-                "Failed to register av_subsystem_meta"
-            );
-        }
-        let meta = json!({
-            "subsystem_type":   subsystem_type_c.as_str(),
-            "hardware_id":      hardware_id_c.as_str(),
-            "confidence_floor": floor,
-        });
-        if let Err(e) = store.save_posture_event_chained(
-            &node_id_c, "AV_ASSET_REGISTERED", &meta.to_string(), None, now,
-        ) {
-            tracing::error!(error=%e, node_id=%node_id_c,
-                "AUDIT-CHAIN WRITE FAILED for AV_ASSET_REGISTERED — event missing from tamper-evident log");
-        }
-    }).await;
-
-    (StatusCode::OK, Json(json!({ "node_id": req.node_id, "registered": true }))).into_response()
-}
 
 // --- Fabric handlers --------------------------------------------------------
 
@@ -2171,213 +635,11 @@ struct RegisterFabricAssetRequest {
     metadata: Option<std::collections::HashMap<String, String>>,
 }
 
-async fn handle_register_fabric_asset(
-    State(svc): State<Arc<ServiceState>>,
-    body: Result<Json<RegisterFabricAssetRequest>, JsonRejection>,
-) -> impl IntoResponse {
-    let req = match body {
-        Ok(Json(r)) => r,
-        Err(e) => return (StatusCode::BAD_REQUEST, Json(json!({"error": e.body_text()}))).into_response(),
-    };
-    if req.asset_id.trim().is_empty() {
-        return (StatusCode::BAD_REQUEST, Json(json!({"error": "asset_id is required"}))).into_response();
-    }
-    let asset = FabricAsset {
-        asset_id: req.asset_id.clone(),
-        asset_type: req.asset_type,
-        display_name: req.display_name,
-        kinematic_profile: req.kinematic_profile,
-        registered_at_ms: now_ms(),
-        last_seen_ms: now_ms(),
-        metadata: req.metadata.unwrap_or_default(),
-    };
-    svc.fabric_router.register_asset(&asset);
-    // #88: if this IS the configured local asset, override the Degraded seed
-    // with fail-closed LockedOut (the feed lifts it); a no-op for peers.
-    seed_local_asset_lockedout(&svc, &req.asset_id);
-    svc.app.store.with(|store| {
-        let _ = store.save_fabric_asset(&asset);
-    });
-    (StatusCode::CREATED, Json(json!({"asset_id": req.asset_id, "registered": true}))).into_response()
-}
 
-async fn handle_list_fabric_assets(
-    State(svc): State<Arc<ServiceState>>,
-) -> impl IntoResponse {
-    let assets = svc.fabric_router.list_assets();
-    let total = assets.len();
-    Json(json!({"assets": assets, "total": total})).into_response()
-}
 
-async fn handle_fabric_state(
-    State(svc): State<Arc<ServiceState>>,
-) -> impl IntoResponse {
-    // SG-007: propagate AND record each rule-firing to the causal log (decisions
-    // unchanged). Use the current fabric generation for the recorded events.
-    let fabric_generation = svc.fabric_router.fabric_state().fabric_generation;
-    let changes = svc.fabric_router.propagate_and_record(&svc.fabric_causal_log, fabric_generation);
-    for (asset_id, new_posture) in changes {
-        let gen = svc.fabric_router.fabric_state().fabric_generation + 1;
-        svc.fabric_router.update_asset_posture(&asset_id, AssetPosture {
-            asset_id: asset_id.clone(),
-            posture: new_posture,
-            generation: gen,
-            computed_at_ms: now_ms(),
-            contributing_nodes: vec![],
-            blocked_by: vec!["cross_asset_propagation".to_string()],
-        });
-    }
-    let state = svc.fabric_router.fabric_state();
-    Json(state).into_response()
-}
 
-async fn handle_fabric_telemetry(
-    State(svc): State<Arc<ServiceState>>,
-) -> impl IntoResponse {
-    let summary = svc.fabric_telemetry.summary();
-    Json(summary).into_response()
-}
 
-async fn handle_fabric_telemetry_asset(
-    State(svc): State<Arc<ServiceState>>,
-    Path(asset_id): Path<String>,
-) -> impl IntoResponse {
-    match svc.fabric_telemetry.asset_snapshot(&asset_id) {
-        Some(snap) => Json(snap).into_response(),
-        None => (StatusCode::NOT_FOUND, Json(json!({"error": "asset not found"}))).into_response(),
-    }
-}
 
-async fn handle_fabric_command(
-    State(svc): State<Arc<ServiceState>>,
-    Path(asset_id): Path<String>,
-    body: Result<Json<ProposedVehicleCommand>, JsonRejection>,
-) -> impl IntoResponse {
-    let cmd = match body {
-        Ok(Json(c)) => c,
-        Err(e) => return (StatusCode::BAD_REQUEST, Json(json!({"error": e.body_text()}))).into_response(),
-    };
-
-    // KIRRA-OCCY-PMON-002: resolve the perception-derate cap O(1) here (the
-    // handler holds `svc`/`ServiceState`) and thread it through to the fabric
-    // governor's Nominal arm. `None` while the monitor is disabled (default).
-    let perception_cap = kirra_verifier::gateway::perception_monitor::resolve_perception_cap(
-        svc.perception_monitor_enabled,
-        &svc.perception_cap,
-        now_ms(),
-    );
-    match svc.fabric_router.route_command(&asset_id, &cmd, perception_cap) {
-        Ok(action) => {
-            use kirra_verifier::gateway::kinematics_contract::EnforceAction;
-            let action_str = format!("{:?}", action);
-            let now = now_ms();
-            let fabric_generation = svc.fabric_router.fabric_state().fabric_generation;
-            let clamp_occurred =
-                matches!(action, EnforceAction::ClampLinear(_) | EnforceAction::ClampSteering(_));
-
-            // #86: APPLY the verdict server-side and return the ENFORCED command.
-            // `apply_enforce_action` substitutes the safe clamp value(s); a clamp
-            // therefore lands in the response's `command` field, so a client using
-            // it is within envelope even if it ignores the `action` label.
-            //
-            // FAIL-CLOSED: deny when no enforced command can be produced —
-            // `DenyBreach`, OR (defensively) a clamp whose enforced value is
-            // non-finite. We NEVER return the unclamped command.
-            let enforced = kirra_verifier::kinematics_sim::apply_enforce_action(&cmd, &action)
-                .filter(|c| c.linear_velocity_mps.is_finite() && c.steering_angle_deg.is_finite());
-
-            match enforced {
-                None => {
-                    // `DenyCode -> &'static str` keeps this path alloc-free; the previous
-                    // `r.clone()` of a `String` allocated per denial (S3 / #115).
-                    let denial_reason: &'static str = match &action {
-                        EnforceAction::DenyBreach(c) => c.reason(),
-                        // Clamp produced a non-finite enforced value (contract bug):
-                        // fail closed rather than forward an invalid command.
-                        _ => "ENFORCED_COMMAND_UNPRODUCIBLE",
-                    };
-                    svc.fabric_causal_log.record(
-                        &asset_id,
-                        "COMMAND_DENIED",
-                        &json!({"reason": denial_reason, "command": serde_json::to_value(&cmd).unwrap_or_default()}).to_string(),
-                        vec![],
-                        vec![],
-                        fabric_generation,
-                    );
-                    // P1: durable audit write off the worker pool (own the asset id,
-                    // materialize the payload).
-                    let asset_id_owned = asset_id.to_string();
-                    let denied_payload =
-                        json!({"asset_id": asset_id, "action": action_str}).to_string();
-                    let _ = svc.app.store.call(move |store| {
-                        let _ = store.save_posture_event_chained(
-                            &asset_id_owned, "FABRIC_COMMAND_DENIED",
-                            &denied_payload,
-                            None, now,
-                        );
-                    }).await;
-                    Json(json!({
-                        "asset_id": asset_id,
-                        "action": action_str,
-                        "allowed": false,
-                        "clamp_occurred": false,
-                        "denial_reason": denial_reason,
-                    })).into_response()
-                }
-                Some(enforced_cmd) => {
-                    // A clamp is safety ENFORCEMENT, not a silent pass: record it
-                    // (causal log + tamper-evident audit) with the original-vs-enforced
-                    // values, mirroring the deny path and the actuator-handler pattern.
-                    if clamp_occurred {
-                        let enforcement = json!({
-                            "asset_id": asset_id,
-                            "action": action_str,
-                            "clamp_occurred": true,
-                            "original_linear_velocity_mps": cmd.linear_velocity_mps,
-                            "original_steering_angle_deg": cmd.steering_angle_deg,
-                            "enforced_linear_velocity_mps": enforced_cmd.linear_velocity_mps,
-                            "enforced_steering_angle_deg": enforced_cmd.steering_angle_deg,
-                        });
-                        svc.fabric_causal_log.record(
-                            &asset_id,
-                            "COMMAND_CLAMPED",
-                            &enforcement.to_string(),
-                            vec![],
-                            vec![],
-                            fabric_generation,
-                        );
-                        // P1: durable audit write off the worker pool.
-                        let asset_id_owned = asset_id.to_string();
-                        let enforcement_str = enforcement.to_string();
-                        let _ = svc.app.store.call(move |store| {
-                            let _ = store.save_posture_event_chained(
-                                &asset_id_owned, "FABRIC_COMMAND_CLAMPED",
-                                &enforcement_str,
-                                None, now,
-                            );
-                        }).await;
-                    }
-
-                    Json(json!({
-                        "asset_id": asset_id,
-                        "action": action_str,
-                        "allowed": true,
-                        "clamp_occurred": clamp_occurred,
-                        "original_linear_velocity_mps": cmd.linear_velocity_mps,
-                        "original_steering_angle_deg": cmd.steering_angle_deg,
-                        "enforced_linear_velocity_mps": enforced_cmd.linear_velocity_mps,
-                        "enforced_steering_angle_deg": enforced_cmd.steering_angle_deg,
-                        // AUTHORITATIVE output: the enforced (post-clamp) command.
-                        "command": enforced_cmd,
-                    })).into_response()
-                }
-            }
-        }
-        Err(e) => {
-            (StatusCode::NOT_FOUND, Json(json!({"error": e.to_string()}))).into_response()
-        }
-    }
-}
 
 #[derive(Deserialize)]
 struct CausalLogQuery {
@@ -2387,61 +649,8 @@ struct CausalLogQuery {
     offset: Option<u32>,
 }
 
-async fn handle_fabric_causal_log(
-    State(svc): State<Arc<ServiceState>>,
-    Query(q): Query<CausalLogQuery>,
-) -> impl IntoResponse {
-    let from = q.from_ms.unwrap_or(0);
-    let to = q.to_ms.unwrap_or(u64::MAX);
-    // #87: bounded + paginated. `limit` is clamped to CAUSAL_EXPORT_MAX_PAGE
-    // inside export_page so a forensic export is never unbounded.
-    let limit = q
-        .limit
-        .unwrap_or(kirra_verifier::fabric::causal_log::CAUSAL_EXPORT_MAX_PAGE);
-    let offset = q.offset.unwrap_or(0);
-    let entries = svc.fabric_causal_log.export_page(from, to, limit, offset);
-    let total = entries.len();
-    Json(json!({"entries": entries, "total": total, "limit": limit, "offset": offset})).into_response()
-}
 
-/// #87: admin-gated verification of the causal-log forensic chain. Mirrors
-/// `/system/audit/verify`. Mounted at `/system/audit/causal/verify` (NOT under
-/// `/fabric/causal-log/...`, to avoid colliding with the `{entry_id}` wildcard).
-async fn verify_causal_chain(
-    State(svc): State<Arc<ServiceState>>,
-) -> impl IntoResponse {
-    let vk = svc.audit_verifying_key.as_ref();
-    match svc.app.store.with(|store| store.verify_causal_chain_integrity(vk)) {
-        Ok(r) => Json(json!({
-            "chain_intact": r.chain_intact,
-            "total_entries": r.total_entries,
-            "latest_hash": r.latest_hash,
-            "signing_enabled": r.signing_enabled,
-            "signed_entries": r.signed_entries,
-            "unsigned_entries": r.unsigned_entries,
-            "signature_valid": r.signature_valid,
-            "first_signed_at_ms": r.first_signed_at_ms,
-            "public_key_b64": r.public_key_b64,
-            "head_verified": r.head_verified,
-            "head_status": r.head_status,
-            "verified": r.chain_intact && r.signature_valid && r.head_verified,
-        })).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "causal chain query failed" }))).into_response(),
-    }
-}
 
-async fn handle_fabric_causal_chain(
-    State(svc): State<Arc<ServiceState>>,
-    Path(entry_id): Path<String>,
-) -> impl IntoResponse {
-    let chain = svc.fabric_causal_log.causal_chain(&entry_id);
-    if chain.is_empty() {
-        return (StatusCode::NOT_FOUND, Json(json!({"error": "entry not found"}))).into_response();
-    }
-    let depth = chain.len();
-    Json(json!({"entry_id": entry_id, "chain": chain, "depth": depth})).into_response()
-}
 
 // ---------------------------------------------------------------------------
 // #46 — systemd `Type=notify` integration: READY notification + watchdog.
@@ -3062,51 +1271,7 @@ async fn main() {
 // gated by the supervisor key below.
 // ===========================================================================
 
-/// GET /console — the operator console UI. One self-contained static file
-/// embedded at build time (`include_str!`): inline CSS+JS, no CDN, no build
-/// step — it works air-gapped.
-async fn console_html() -> impl IntoResponse {
-    Html(include_str!("../../static/console.html"))
-}
 
-/// GET /console/fleet — per-node posture summary from the durable store
-/// (`load_nodes`). QM read. `note` is the Untrusted reason carried on the node's
-/// trust state (the latest trust note); `null` for Trusted/Unknown.
-async fn console_fleet(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    // load_nodes + the per-node clearance lookups run under ONE acquisition
-    // (Rule 5). The closure returns a Result; the error response is built outside.
-    let fleet = svc.app.store.with(|store| {
-        let nodes = store.load_nodes().map_err(|_| "load_nodes failed")?;
-        let fleet: Vec<_> = nodes
-            .iter()
-            .map(|n| {
-                let (posture, note) = match &n.status {
-                    NodeTrustState::Trusted => ("Trusted", None),
-                    NodeTrustState::Untrusted(reason) => ("Untrusted", Some(reason.clone())),
-                    NodeTrustState::Unknown => ("Unknown", None),
-                };
-                // Phase B: the latest clearance grant's delivery state (or null). The
-                // UI derives the lifecycle label (pending / delivered:Cleared /
-                // delivery-rejected:reason) from these raw columns — no invented state.
-                let clearance = store.latest_clearance_grant(&n.node_id).ok().flatten();
-                json!({
-                    "node_id": n.node_id,
-                    "posture": posture,
-                    "note": note,
-                    "last_seen_ms": n.last_trust_update_ms,
-                    "clearance": clearance,
-                })
-            })
-            .collect();
-        Ok::<_, &'static str>(fleet)
-    });
-    let fleet = match fleet {
-        Ok(f) => f,
-        Err(msg) => return (StatusCode::INTERNAL_SERVER_ERROR,
-                            Json(json!({ "error": msg }))).into_response(),
-    };
-    Json(json!({ "fleet": fleet, "total": fleet.len() })).into_response()
-}
 
 #[derive(Deserialize)]
 struct ConsoleAuditQuery {
@@ -3114,130 +1279,8 @@ struct ConsoleAuditQuery {
     offset: Option<u64>,
 }
 
-/// GET /console/audit?limit=&offset= — passthrough to `load_audit_chain_page`.
-///
-/// GROUNDING NOTE: `load_audit_chain_page` is **offset-paginated** (DESC by id),
-/// not before-seq cursored. The console pages by offset (the "load older" control
-/// increments `offset` over the DESC feed — the same backward paging the original
-/// `?before=<seq>` intent describes). Returns exactly what the page rows carry
-/// (sequence, event_type, payload, signature key-id, per-row signature status,
-/// and the page-level `chain_intact` flag) — no invented fields. QM read.
-async fn console_audit(
-    State(svc): State<Arc<ServiceState>>,
-    Query(params): Query<ConsoleAuditQuery>,
-) -> impl IntoResponse {
-    let limit = params.limit.unwrap_or(50).min(500);
-    let offset = params.offset.unwrap_or(0);
-    let vk = svc.audit_verifying_key.as_ref();
-    match svc.app.store.with(|store| store.load_audit_chain_page(limit, offset, vk)) {
-        Ok(page) => Json(page).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "audit query failed" }))).into_response(),
-    }
-}
 
-/// GET /console/escalations — derived view of OPEN SG6 escalations.
-///
-/// CONSERVATIVE SUPERSET + AMBIGUITY NOTE: the SG6 impact lifecycle events
-/// (`ImpactDetected` / `ImpactEscalationRaised` / `ImpactCleared`, parko-kirra
-/// `audit_sink`, #102/#103) DO land in this audit chain — but their payloads
-/// (`ImpactDetectedPayload`) carry **no node/asset id**, so per-node attribution
-/// is NOT derivable from the chain alone. This view therefore returns the
-/// fleet-level OPEN superset over a single timeline: scanning most-recent-first,
-/// the detect/raise events seen before the first `ImpactCleared` are "open"
-/// (events since the last clear). It passes through whatever each event row
-/// carries — no fabricated node_id. Per-node attribution is a Phase-B / event-
-/// taxonomy enhancement. QM read.
-async fn console_escalations(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    let vk = svc.audit_verifying_key.as_ref();
-    let page = match svc.app.store.with(|store| store.load_audit_chain_page(1000, 0, vk)) {
-        Ok(p) => p,
-        Err(_) => {
-            return (StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(json!({ "error": "audit query failed" }))).into_response()
-        }
-    };
-    let mut open = Vec::new();
-    for e in &page.entries {
-        let v = serde_json::to_value(e).unwrap_or(serde_json::Value::Null);
-        let et = v.get("event_type").and_then(|x| x.as_str()).unwrap_or("");
-        if et == "ImpactCleared" {
-            break; // most-recent clear closes the single-timeline superset
-        }
-        if et == "ImpactDetected" || et == "ImpactEscalationRaised" {
-            open.push(v);
-        }
-    }
-    Json(json!({
-        "open_escalations": open,
-        "count": open.len(),
-        "note": "fleet-level superset — impact events carry no node id (see handler doc); attribution is Phase B",
-    }))
-    .into_response()
-}
 
-/// GET /console/runtime (#395) — public read-only runtime snapshot.
-///
-/// Composes live in-memory state (mode, uptime, generation, cache freshness,
-/// node/asset counts, broadcast fanout, HA heartbeat age) with two store reads
-/// (audit chain depth + the persisted heartbeat ms). Fail-closed: store-lock
-/// poison or query error → 500 json error (mirrors `console_audit`).
-async fn console_runtime(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    let now = now_ms();
-
-    // last_recalc_ms comes from the atomic posture-cache snapshot (0 if cold).
-    // `SharedPostureCache` is a std `RwLock`; a poisoned lock fails closed → 500.
-    let last_recalc_ms = match svc.posture_cache.read() {
-        Ok(guard) => guard.as_ref().map(|c| c.generated_at_ms).unwrap_or(0),
-        Err(_) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": "posture cache lock poisoned" })),
-            )
-                .into_response()
-        }
-    };
-
-    // Two store reads under one lock acquisition: audit depth + HA heartbeat.
-    // The closure returns a Result; per-read error responses are built outside.
-    let probe = svc.app.store.with(|store| {
-        let audit_entries = store.audit_chain_len().map_err(|_| "audit query failed")?;
-        // Heartbeat absent → null (no primary has written yet).
-        let hb = store.load_engine_state(HEARTBEAT_KEY)
-            .map_err(|_| "engine state query failed")?
-            .and_then(|v| v.parse::<u64>().ok())
-            .map(|stored| now.saturating_sub(stored));
-        Ok::<_, &'static str>((audit_entries, hb))
-    });
-    let (audit_entries, ha_heartbeat_age_ms) = match probe {
-        Ok(pair) => pair,
-        Err(msg) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": msg })),
-            )
-                .into_response()
-        }
-    };
-
-    let mode = if svc.app.is_active() { "Active" } else { "PassiveStandby" };
-
-    Json(json!({
-        "mode": mode,
-        "uptime_ms": now.saturating_sub(svc.started_at_ms),
-        "posture_generation": kirra_verifier::posture_engine::POSTURE_GENERATION
-            .load(std::sync::atomic::Ordering::SeqCst),
-        "last_recalc_ms": last_recalc_ms,
-        "posture_cache_ttl_ms": POSTURE_CACHE_TTL_MS,
-        "total_nodes": svc.app.nodes.len(),
-        "fabric_assets": svc.fabric_router.fabric_state().total_assets,
-        "fabric_denial_rate": svc.fabric_telemetry.summary().fabric_denial_rate,
-        "audit_entries": audit_entries,
-        "broadcast_subscribers": svc.app.posture_tx.receiver_count(),
-        "ha_heartbeat_age_ms": ha_heartbeat_age_ms,
-    }))
-    .into_response()
-}
 
 /// Query for #396 console analytics. `window_ms` defaults to 24h.
 #[derive(Deserialize)]
@@ -3246,207 +1289,8 @@ struct ConsoleAnalyticsQuery {
     window_ms: Option<u64>,
 }
 
-/// GET /console/analytics?window_ms= (#396) — public read-only time-series view.
-///
-/// Buckets EXISTING `posture_events` rows over `[since_ms, now]` into 24 buckets
-/// (no new data class). Fail-closed: store-lock poison / query error → 500.
-///
-/// DATA-AVAILABILITY NOTES (honest about what is and isn't stored):
-///   - `posture_transitions`: real — bucketed from `posture_events.posture_json`
-///     (the resulting `FleetPosture`).
-///   - `denial_rate_series`: the store keeps NO per-bucket denial history. We
-///     therefore emit a SINGLE current-value point (the live fabric denial rate)
-///     rather than fabricate a fake series — the array shape is preserved.
-///   - `interventions_by_asset`: fabric telemetry tracks a per-asset `denial_rate`
-///     and `commands_per_minute`, NOT separate clamp/deny COUNTERS. `denies` is
-///     derived (rate × cpm, rounded); `clamps` is 0 — clamp counts are not stored.
-///   - `flapping_top`: real — per-node posture-event counts since `since_ms`.
-async fn console_analytics(
-    State(svc): State<Arc<ServiceState>>,
-    Query(params): Query<ConsoleAnalyticsQuery>,
-) -> impl IntoResponse {
-    const BUCKETS: u64 = 24;
-    const FLAPPING_TOP_N: usize = 10;
 
-    let now = now_ms();
-    let window_ms = params.window_ms.unwrap_or(86_400_000).max(1);
-    let since_ms = now.saturating_sub(window_ms);
-    let bucket_span = (window_ms / BUCKETS).max(1);
 
-    // Both reads share ONE acquisition (Rule 5); the closure returns a Result and
-    // the error response is produced outside (Rule 4).
-    let loaded = svc.app.store.with(|store| {
-        let events = store.load_posture_events_since(since_ms)
-            .map_err(|_| "posture event query failed")?;
-        let by_node = store.count_posture_events_by_node_since(since_ms)
-            .map_err(|_| "posture event query failed")?;
-        Ok::<_, &'static str>((events, by_node))
-    });
-    let (events, by_node) = match loaded {
-        Ok(pair) => pair,
-        Err(msg) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({ "error": msg })),
-            )
-                .into_response()
-        }
-    };
-
-    // Bucket posture transitions by resulting posture (parsed from posture_json).
-    let mut to_degraded = vec![0u64; BUCKETS as usize];
-    let mut to_lockedout = vec![0u64; BUCKETS as usize];
-    let mut to_nominal = vec![0u64; BUCKETS as usize];
-    for (created_at_ms, posture_json) in &events {
-        let idx = (created_at_ms.saturating_sub(since_ms) / bucket_span)
-            .min(BUCKETS - 1) as usize;
-        // posture_json serializes FleetPosture as "Nominal"/"Degraded"/"LockedOut"
-        // (or {"Untrusted": "..."} style for node states). Match the variant name.
-        let v: serde_json::Value =
-            serde_json::from_str(posture_json).unwrap_or(serde_json::Value::Null);
-        let label = v.as_str().map(|s| s.to_string()).unwrap_or_else(|| {
-            // Object-tagged form: take the first key.
-            v.as_object()
-                .and_then(|o| o.keys().next().cloned())
-                .unwrap_or_default()
-        });
-        match label.as_str() {
-            "Degraded" => to_degraded[idx] += 1,
-            "LockedOut" => to_lockedout[idx] += 1,
-            "Nominal" => to_nominal[idx] += 1,
-            _ => {}
-        }
-    }
-    let posture_transitions: Vec<serde_json::Value> = (0..BUCKETS as usize)
-        .map(|i| {
-            json!({
-                "bucket_start_ms": since_ms + (i as u64) * bucket_span,
-                "to_degraded": to_degraded[i],
-                "to_lockedout": to_lockedout[i],
-                "to_nominal": to_nominal[i],
-            })
-        })
-        .collect();
-
-    // denial_rate_series: no stored per-bucket history → single current point.
-    let denial_rate_series = json!([{
-        "bucket_start_ms": now,
-        "denial_rate": svc.fabric_telemetry.summary().fabric_denial_rate,
-    }]);
-
-    // interventions_by_asset: derive denies from rate × cpm; clamps not stored.
-    let interventions_by_asset: Vec<serde_json::Value> = svc
-        .fabric_telemetry
-        .all_snapshots()
-        .into_iter()
-        .map(|s| {
-            let denies = (s.denial_rate * s.commands_per_minute).round() as u64;
-            json!({
-                "asset_id": s.asset_id,
-                "clamps": 0, // clamp counts are not separately tracked (see doc)
-                "denies": denies,
-            })
-        })
-        .collect();
-
-    let flapping_top: Vec<serde_json::Value> = by_node
-        .into_iter()
-        .take(FLAPPING_TOP_N)
-        .map(|(node_id, transitions)| json!({
-            "node_id": node_id,
-            "transitions": transitions,
-        }))
-        .collect();
-
-    Json(json!({
-        "window_ms": window_ms,
-        "posture_transitions": posture_transitions,
-        "denial_rate_series": denial_rate_series,
-        "interventions_by_asset": interventions_by_asset,
-        "flapping_top": flapping_top,
-    }))
-    .into_response()
-}
-
-/// GET /console/sites (#397) — public read-only site rollup over in-memory nodes.
-///
-/// MAPPING CHOICE (documented): the rollup is by node TRUST STATUS, not DAG
-/// fleet posture. `NodeTrustState::Trusted` → nominal; `Unknown` → degraded
-/// bucket; `Untrusted(_)` → lockedout bucket. Nodes with a NULL `site` roll into
-/// `unassigned`. Pure in-memory read (no store lock); cannot fail-closed.
-async fn console_sites(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    use std::collections::BTreeMap;
-    // (total, nominal, degraded, lockedout)
-    let mut sites: BTreeMap<String, (u64, u64, u64, u64)> = BTreeMap::new();
-    let mut unassigned: u64 = 0;
-
-    for entry in svc.app.nodes.iter() {
-        let node = entry.value();
-        let bucket = match &node.status {
-            NodeTrustState::Trusted => 0,
-            NodeTrustState::Unknown => 1,
-            NodeTrustState::Untrusted(_) => 2,
-        };
-        match &node.site {
-            Some(site) => {
-                let e = sites.entry(site.clone()).or_insert((0, 0, 0, 0));
-                e.0 += 1;
-                match bucket {
-                    0 => e.1 += 1,
-                    1 => e.2 += 1,
-                    _ => e.3 += 1,
-                }
-            }
-            None => unassigned += 1,
-        }
-    }
-
-    let sites: Vec<serde_json::Value> = sites
-        .into_iter()
-        .map(|(site, (total, nominal, degraded, lockedout))| json!({
-            "site": site,
-            "total": total,
-            "nominal": nominal,
-            "degraded": degraded,
-            "lockedout": lockedout,
-        }))
-        .collect();
-
-    Json(json!({ "sites": sites, "unassigned": unassigned })).into_response()
-}
-
-/// GET /console/versions (#398) — public read-only firmware-version rollup over
-/// in-memory nodes. Nodes with a NULL `firmware_version` count toward `unknown`.
-/// `pct` = count/total*100 (guarded against divide-by-zero). Pure in-memory read.
-async fn console_versions(State(svc): State<Arc<ServiceState>>) -> impl IntoResponse {
-    use std::collections::BTreeMap;
-    let mut versions: BTreeMap<String, u64> = BTreeMap::new();
-    let mut unknown: u64 = 0;
-    let mut total: u64 = 0;
-
-    for entry in svc.app.nodes.iter() {
-        total += 1;
-        match &entry.value().firmware_version {
-            Some(v) => *versions.entry(v.clone()).or_insert(0) += 1,
-            None => unknown += 1,
-        }
-    }
-
-    let versions: Vec<serde_json::Value> = versions
-        .into_iter()
-        .map(|(version, count)| {
-            let pct = if total > 0 {
-                (count as f64) / (total as f64) * 100.0
-            } else {
-                0.0
-            };
-            json!({ "version": version, "count": count, "pct": pct })
-        })
-        .collect();
-
-    Json(json!({ "versions": versions, "total": total, "unknown": unknown }))
-        .into_response()
-}
 
 /// Pure supervisor-key decision (testable without env — INV-13 forbids `set_var`
 /// in multithreaded tests). REUSES the #255 mechanism: the value is the
@@ -3534,112 +1378,7 @@ struct RegisterOperatorRequest {
     ed25519_pubkey_pem: String,
 }
 
-/// POST /console/operators — register (or rotate) an operator's Ed25519 PUBLIC key
-/// (#314 Phase 1). **ADMIN-token-gated** (the node-registration precedent). Admin
-/// and supervisor are SEPARATE powers: this route lives behind
-/// `require_admin_token`, so the **supervisor key cannot register operators**. The
-/// private key NEVER reaches the server — only the public key is stored.
-async fn register_operator(
-    State(svc): State<Arc<ServiceState>>,
-    headers: HeaderMap,
-    Json(req): Json<RegisterOperatorRequest>,
-) -> impl IntoResponse {
-    let operator_id = req.operator_id.trim();
-    if operator_id.is_empty() {
-        return (StatusCode::UNPROCESSABLE_ENTITY,
-                Json(json!({ "error": "operator_id must be non-empty" }))).into_response();
-    }
-    // #326: reject the legacy `|` delimiter and control characters in the id.
-    if !valid_identifier(operator_id) {
-        return (StatusCode::BAD_REQUEST, Json(json!({
-            "error": "operator_id must not contain '|' or control characters"
-        }))).into_response();
-    }
-    // Fail-closed: the PEM must parse as a valid Ed25519 SPKI public key.
-    let fingerprint = match kirra_verifier::attestation::operator_key_fingerprint(
-        &req.ed25519_pubkey_pem,
-    ) {
-        Some(fp) => fp,
-        None => return (StatusCode::UNPROCESSABLE_ENTITY, Json(json!({
-            "error": "ed25519_pubkey_pem is not a valid Ed25519 SubjectPublicKeyInfo PEM"
-        }))).into_response(),
-    };
-    let now = now_ms();
-    // The revoked-check + register + audit run under ONE acquisition (Rule 5).
-    // The closure returns the fully-built response (the early persist-failure
-    // path returns its 500 response from inside the closure — Rule 4).
-    svc.app.store.with(|store| {
-        // #327: detect a prior REVOKED row BEFORE registering — register_operator
-        // silently clears revoked_at, so reactivation would otherwise be
-        // invisible in the ledger. Record it as a distinct, attributed event.
-        let was_revoked = store
-            .load_operator(operator_id)
-            .ok()
-            .flatten()
-            .is_some_and(|o| o.revoked_at_ms.is_some());
-        if store.register_operator(operator_id, &req.ed25519_pubkey_pem, now).is_err() {
-            return (StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(json!({ "error": "persist failed" }))).into_response();
-        }
-        if was_revoked {
-            // A previously-revoked operator is now active again — the
-            // reactivating admin is attributed by token fingerprint (#327).
-            let _ = store.append_clearance_audit_event(
-                "OperatorReactivated",
-                &json!({
-                    "operator_id": operator_id,
-                    "operator_key_fingerprint": fingerprint,
-                    "reactivated_by_admin_fingerprint": admin_token_fingerprint(&headers),
-                }).to_string(),
-                now,
-            );
-            (StatusCode::CREATED, Json(json!({
-                "operator_id": operator_id,
-                "operator_key_fingerprint": fingerprint,
-                "status": "reactivated",
-            }))).into_response()
-        } else {
-            let _ = store.append_clearance_audit_event(
-                "OperatorRegistered",
-                &json!({ "operator_id": operator_id, "operator_key_fingerprint": fingerprint })
-                    .to_string(),
-                now,
-            );
-            (StatusCode::CREATED, Json(json!({
-                "operator_id": operator_id,
-                "operator_key_fingerprint": fingerprint,
-                "status": "registered",
-            }))).into_response()
-        }
-    })
-}
 
-/// POST /console/operators/{operator_id}/revoke — revoke an operator (#314).
-/// ADMIN-token-gated. A revoked operator can never clear a grant.
-async fn revoke_operator(
-    State(svc): State<Arc<ServiceState>>,
-    Path(operator_id): Path<String>,
-) -> impl IntoResponse {
-    let now = now_ms();
-    // revoke + its audit event share one acquisition (Rule 5); the closure
-    // returns the built response.
-    svc.app.store.with(|store| match store.revoke_operator(&operator_id, now) {
-        Ok(true) => {
-            let _ = store.append_clearance_audit_event(
-                "OperatorRevoked",
-                &json!({ "operator_id": operator_id }).to_string(),
-                now,
-            );
-            (StatusCode::OK, Json(json!({ "operator_id": operator_id, "status": "revoked" })))
-                .into_response()
-        }
-        Ok(false) => (StatusCode::NOT_FOUND,
-                      Json(json!({ "error": "operator not found or already revoked" })))
-            .into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "error": "persist failed" }))).into_response(),
-    })
-}
 
 #[derive(Deserialize)]
 struct ClearanceChallengeQuery {
@@ -3647,51 +1386,6 @@ struct ClearanceChallengeQuery {
     node_id: String,
 }
 
-/// GET /console/clearance-challenge?operator_id=&node_id= — issue a one-time nonce
-/// the operator signs to prove key possession (#314 Phase 1; the attestation
-/// challenge-issuance pattern). Unauthenticated (the nonce alone grants nothing —
-/// only a valid signature over it does). #325: NO enumeration oracle — every
-/// well-formed request returns a uniform 200 with a nonce-shaped body, so an
-/// unknown/revoked operator is indistinguishable from a known one. A real challenge
-/// is stored ONLY for an ACTIVE operator (verify-then-consume at grant time); anyone
-/// else gets a DECOY nonce that is never stored (no map growth) and that a grant
-/// attempt still rejects at the unchanged grant-time operator check.
-async fn clearance_challenge(
-    State(svc): State<Arc<ServiceState>>,
-    Query(q): Query<ClearanceChallengeQuery>,
-) -> impl IntoResponse {
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    let operator_id = q.operator_id.trim();
-    let node_id = q.node_id.trim();
-    if operator_id.is_empty() || node_id.is_empty() {
-        return (StatusCode::UNPROCESSABLE_ENTITY,
-                Json(json!({ "error": "operator_id and node_id are required" }))).into_response();
-    }
-    let active = svc.app.store.with(|store| store
-        .load_operator(operator_id)
-        .ok()
-        .flatten()
-        .map(|o| o.is_active())
-        .unwrap_or(false));
-    // Hex string (not a u64) so the in-browser signing flow never loses precision.
-    let nonce_hex = format!("{:016x}", kirra_verifier::verifier::generate_challenge_nonce());
-    // #325: store a REAL challenge only for an active operator; everyone else gets a
-    // decoy nonce that is never recorded (no oracle, no map growth). #326: the
-    // challenge-map key is length-prefixed (unambiguous operator/node split).
-    if active {
-        let key = composite_challenge_key(operator_id, node_id);
-        svc.app.issue_clearance_challenge(&key, nonce_hex.clone(), now_ms());
-    }
-    (StatusCode::OK, Json(json!({
-        "operator_id": operator_id,
-        "node_id": node_id,
-        "nonce": nonce_hex,
-        "ttl_ms": 30_000,
-    }))).into_response()
-}
 
 #[derive(Deserialize)]
 struct ClearanceGrantRequest {
@@ -3705,156 +1399,6 @@ struct ClearanceGrantRequest {
     signature_b64: Option<String>,
 }
 
-/// POST /console/clearance-grants — the ONE inbound console affordance, upgraded
-/// to operator-proven identity (#314 Phase 1).
-///
-/// **RECORD-ONLY (Phase A):** records + signs a clearance grant; it does NOT
-/// release the node (delivery to the node `ClearanceLoop` is Phase B). The server
-/// stamps `granted_at_ms` from ITS clock (no client time → no future-dating).
-/// Never mutates posture.
-///
-/// Two auth paths:
-///   * **operator-signed (PRIMARY):** a registered operator proves key possession
-///     by signing the challenge nonce. Handler order mirrors `verify_attestation`:
-///     load operator (unknown/revoked → 403) → VERIFY signature → CONSUME nonce
-///     (verify-then-consume; replay → 401) → well-formedness → record with
-///     `auth_method="operator-signed"` + the key fingerprint, BOTH embedded in the
-///     signed chain event (the non-repudiation payoff).
-///   * **supervisor-break-glass (NAMED FALLBACK):** the shared `KIRRA_SUPERVISOR_
-///     RESET_KEY` (#255) REMAINS as an explicitly-named break-glass path —
-///     `auth_method="supervisor-break-glass"`, audited distinctly and loudly —
-///     because a fleet locked out of recovery by a lost operator key is its OWN
-///     safety failure. Deployments disable it by unsetting the env. The console UI
-///     presents operator-signed as the primary flow.
-async fn console_clearance_grant(
-    State(svc): State<Arc<ServiceState>>,
-    headers: HeaderMap,
-    Json(req): Json<ClearanceGrantRequest>,
-) -> impl IntoResponse {
-    // HA split-brain guard (#323): recording a grant is a store MUTATION. The
-    // `/console` posture-exemption keeps this reachable during LockedOut, but that
-    // is about POSTURE — it does NOT justify a passive-standby instance writing
-    // grants and diverging from the primary. Fail-closed like every other mutation.
-    if !svc.app.is_active() {
-        return (StatusCode::SERVICE_UNAVAILABLE,
-                Json(json!({ "error": "instance is in passive standby mode" }))).into_response();
-    }
-    let now = now_ms();
-    let operator_id = req.operator_id.trim().to_string();
-    let node_id = req.node_id.trim().to_string();
-
-    let operator_signed = req.signature_b64.as_deref().is_some_and(|s| !s.is_empty());
-
-    let (auth_method, fingerprint): (&str, Option<String>) = if operator_signed {
-        // === OPERATOR-SIGNED PATH — verify-then-consume (mirrors verify_attestation).
-        let nonce = req.nonce.clone().unwrap_or_default();
-        let sig_b64 = req.signature_b64.clone().unwrap_or_default();
-        if nonce.is_empty() {
-            audit_grant_rejection(&svc.app, "missing_nonce", &node_id, &operator_id, now);
-            return (StatusCode::UNPROCESSABLE_ENTITY, Json(json!({
-                "status": "rejected", "reason": "nonce required for an operator-signed grant"
-            }))).into_response();
-        }
-        // 1. Load operator — unknown / revoked → 403, audited.
-        let operator = match svc.app.store.with(|s| s.load_operator(&operator_id).ok().flatten()) {
-            Some(op) if op.is_active() => op,
-            Some(_) => {
-                audit_grant_rejection(&svc.app, "revoked_operator", &node_id, &operator_id, now);
-                return (StatusCode::FORBIDDEN, Json(json!({
-                    "status": "rejected", "reason": "operator is revoked"
-                }))).into_response();
-            }
-            None => {
-                audit_grant_rejection(&svc.app, "unknown_operator", &node_id, &operator_id, now);
-                return (StatusCode::FORBIDDEN, Json(json!({
-                    "status": "rejected", "reason": "operator is not registered"
-                }))).into_response();
-            }
-        };
-        // 2. VERIFY signature FIRST (before the nonce is consumed).
-        use base64::{engine::general_purpose::STANDARD as b64e, Engine as _};
-        let sig_bytes = match b64e.decode(sig_b64.trim()) {
-            Ok(b) => b,
-            Err(_) => {
-                audit_grant_rejection(&svc.app, "malformed_signature", &node_id, &operator_id, now);
-                return (StatusCode::UNAUTHORIZED, Json(json!({
-                    "status": "rejected", "reason": "signature is not valid base64"
-                }))).into_response();
-            }
-        };
-        let payload = kirra_verifier::attestation::operator_grant_signing_payload(
-            &operator_id, &node_id, &nonce,
-        );
-        if !kirra_verifier::attestation::verify_ed25519_pem_signature(
-            &operator.pubkey_pem, &payload, &sig_bytes,
-        ) {
-            audit_grant_rejection(&svc.app, "bad_signature", &node_id, &operator_id, now);
-            return (StatusCode::UNAUTHORIZED, Json(json!({
-                "status": "rejected", "reason": "signature verification failed"
-            }))).into_response();
-        }
-        // 3. CONSUME the nonce (verify-then-consume; replay/expired → 401, audited).
-        //    #326: same length-prefixed composite key the challenge issuer used.
-        let key = composite_challenge_key(&operator_id, &node_id);
-        if !svc.app.consume_clearance_challenge(&key, &nonce, now) {
-            audit_grant_rejection(&svc.app, "nonce_replay_or_expired", &node_id, &operator_id, now);
-            return (StatusCode::UNAUTHORIZED, Json(json!({
-                "status": "rejected",
-                "reason": "challenge nonce absent, expired, or already used (replay rejected)"
-            }))).into_response();
-        }
-        let fp = kirra_verifier::attestation::operator_key_fingerprint(&operator.pubkey_pem);
-        ("operator-signed", fp)
-    } else {
-        // === BREAK-GLASS PATH — the named, distinctly-audited supervisor fallback.
-        if let Err(code) = check_supervisor_key(&headers) {
-            if code == StatusCode::UNAUTHORIZED {
-                audit_grant_rejection(&svc.app, "supervisor_unauthorized", &node_id, &operator_id, now);
-            }
-            return (code, Json(json!({
-                "status": "rejected",
-                "reason": "no operator signature and supervisor authorization failed"
-            }))).into_response();
-        }
-        tracing::warn!(node_id = %node_id, operator_id = %operator_id,
-            "BREAK-GLASS clearance grant via supervisor key — operator-signed path bypassed \
-             (auth_method=supervisor-break-glass; audited distinctly)");
-        ("supervisor-break-glass", None)
-    };
-
-    // Shared well-formedness (mirrors OperatorClearanceGrant::is_well_formed).
-    if operator_id.is_empty() {
-        audit_grant_rejection(&svc.app, "empty_operator_id", &node_id, &operator_id, now);
-        return (StatusCode::UNPROCESSABLE_ENTITY, Json(json!({
-            "status": "rejected", "reason": "operator_id must be non-empty"
-        }))).into_response();
-    }
-    let registered = svc.app.store.with(|store| store.node_exists(&node_id).unwrap_or(false));
-    if !registered {
-        audit_grant_rejection(&svc.app, "unregistered_node", &node_id, &operator_id, now);
-        return (StatusCode::UNPROCESSABLE_ENTITY, Json(json!({
-            "status": "rejected", "reason": "node_id is not a registered node"
-        }))).into_response();
-    }
-
-    // Record + sign — RECORD-ONLY. Delivery is Phase B; posture is untouched.
-    svc.app.store.with(|store| match store.save_clearance_grant_chained_with_auth(
-        &node_id, &operator_id, now, auth_method, fingerprint.as_deref(),
-    ) {
-        Ok(_id) => (StatusCode::OK, Json(json!({
-            "status": "recorded",
-            "delivery": "pending-phase-b",
-            "node_id": node_id,
-            "operator_id": operator_id,
-            "granted_at_ms": now,
-            "auth_method": auth_method,
-            "operator_key_fingerprint": fingerprint,
-            "note": "grant recorded and signed; the vehicle is NOT released — delivery to the node ClearanceLoop is Phase B",
-        }))).into_response(),
-        Err(_) => (StatusCode::INTERNAL_SERVER_ERROR,
-                   Json(json!({ "status": "error", "reason": "persist failed" }))).into_response(),
-    })
-}
 
 /// Assembles the complete production router from a fully-initialized
 /// `ServiceState`. Extracted verbatim from `main()` (issue #72) so the EXACT
@@ -6288,12 +3832,24 @@ mod store_offload_guard {
     #[test]
     fn audit_chain_write_is_never_on_a_sync_worker() {
         // Embeds a compile-time snapshot of this very file (path is relative to it).
-        let src = include_str!("kirra_verifier_service.rs");
-
-        let mut nearest_access = ""; // "with" (sync) | "call" (off-worker)
-        let mut in_test = false; // all handlers precede the first test module here
+        let sources: [&str; 11] = [
+            include_str!("kirra_verifier_service.rs"),
+            include_str!("kirra_verifier_service/attestation.rs"),
+            include_str!("kirra_verifier_service/fleet.rs"),
+            include_str!("kirra_verifier_service/audit.rs"),
+            include_str!("kirra_verifier_service/action_filter.rs"),
+            include_str!("kirra_verifier_service/industrial.rs"),
+            include_str!("kirra_verifier_service/federation.rs"),
+            include_str!("kirra_verifier_service/actuator.rs"),
+            include_str!("kirra_verifier_service/fabric.rs"),
+            include_str!("kirra_verifier_service/console.rs"),
+            include_str!("kirra_verifier_service/operators.rs"),
+        ];
         let mut violations: Vec<usize> = Vec::new();
 
+        for src in sources {
+        let mut nearest_access = ""; // "with" (sync) | "call" (off-worker)
+        let mut in_test = false; // tests live only in the root file; submodules are all production
         for (idx, line) in src.lines().enumerate() {
             if line.trim_start().starts_with("#[cfg(test)]") {
                 in_test = true;
@@ -6311,6 +3867,7 @@ mod store_offload_guard {
             {
                 violations.push(idx + 1);
             }
+        }
         }
 
         assert!(
