+++ src/bin/kirra_verifier_service/actuator.rs
@@ -1,0 +1,58 @@
+// src/bin/kirra_verifier_service/actuator.rs
+// actuator route handlers — split from kirra_verifier_service.rs (pure move).
+//
+// `use super::*` pulls the binary root's DTOs, helpers and `use` imports
+// (visible to this descendant module); handlers are `pub(crate)` so the
+// root re-export (`use actuator::*`) lets build_app/tests name them unqualified.
+
+use super::*;
+
+pub(crate) async fn handle_actuator_motion_command(
+    State(svc): State<Arc<ServiceState>>,
+    // Threaded by `enforce_actuator_safety_envelope` (the route always runs it
+    // first). It carries the TRUE verdict — `cmd` below is already the enforced
+    // (post-clamp) command, but only this tells us WHETHER a clamp happened, so
+    // the response can report it instead of always claiming "Allow".
+    Extension(outcome): Extension<EnforcementOutcome>,
+    Json(cmd): Json<ProposedVehicleCommand>,
+) -> impl IntoResponse {
+    let now = now_ms();
+
+    tracing::info!(
+        action              = %outcome.action.as_str(),
+        linear_velocity_mps = %cmd.linear_velocity_mps,
+        steering_angle_deg  = %cmd.steering_angle_deg,
+        delta_time_s        = %cmd.delta_time_s,
+        "Actuator motion command admitted through safety envelope"
+    );
+
+    // Record the verdict in the tamper-evident log, including whether a clamp
+    // occurred and the original vs enforced values (previously a clamp was
+    // logged indistinguishably from a plain admit).
+    let audit = serde_json::json!({
+        "action":                       outcome.action.as_str(),
+        "original_linear_velocity_mps": outcome.original_linear_velocity_mps,
+        "original_steering_angle_deg":  outcome.original_steering_angle_deg,
+        "enforced_linear_velocity_mps": outcome.enforced_linear_velocity_mps,
+        "enforced_steering_angle_deg":  outcome.enforced_steering_angle_deg,
+        "current_velocity_mps":         cmd.current_velocity_mps,
+        "current_steering_angle_deg":   cmd.current_steering_angle_deg,
+        "delta_time_s":                 cmd.delta_time_s,
+        "admitted_at_ms":               now,
+    });
+    // P1: durable audit write off the worker pool (materialize the payload first).
+    let audit_str = audit.to_string();
+    let _ = svc.app.store.call(move |store| {
+        if let Err(e) = store.save_posture_event_chained(
+            "actuator_motion", "MOTION_COMMAND_ADMITTED",
+            &audit_str, None, now,
+        ) {
+            tracing::error!(error=%e,
+                "AUDIT-CHAIN WRITE FAILED for MOTION_COMMAND_ADMITTED — event missing from tamper-evident log");
+        }
+    }).await;
+
+    // Response speaks the ROS interceptor's schema (action / enforced_*) AND
+    // the legacy keys (now accurate). See `EnforcementOutcome::response_body`.
+    (StatusCode::OK, Json(outcome.response_body())).into_response()
+}
